% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{add_labels}
\alias{add_labels}
\alias{remove_labels}
\title{Add or remove labels from an issue or pull request}
\usage{
add_labels(labels, issue, repo, ...)

remove_labels(labels, issue, repo, ...)
}
\arguments{
\item{labels}{(character) The name of the labels.}

\item{issue}{(string or character) The number or title of the issue.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{add_labels()} returns a tibble of the added labels properties. \code{remove_labels()}
returns a tibble of the remaining labels properties.

\strong{Label Properties:}
\itemize{
\item \strong{name}: The name of the label.
\item \strong{color}: The color of the label in hexidecimal code.
\item \strong{description}: The description of the label.
}
}
\description{
\code{add_labels()} adds labels to an existing issue or pull request within a GitHub repository.
\code{remove_labels()} removes labels from an issue or pull request.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/labels/#add-labels-to-an-issue}
\item \url{https://developer.github.com/v3/issues/labels/#remove-a-label-from-an-issue}
}
}
\examples{
\dontrun{

  add_labels(
    labels = c("feature", "new-label"),
    issue  = "Test issue",
    repo   = "ChadGoymer/test-githapi")

  remove_labels(
    labels = c("feature", "new-label"),
    issue  = "Test issue",
    repo   = "ChadGoymer/test-githapi")

}

}
