% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branches.R
\name{branch_exists}
\alias{branch_exists}
\title{Determine whether a branch exists in the specified repository.}
\usage{
branch_exists(repo, branch, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{branch}{(character) The name pf the branch.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function returns \code{TRUE} if the branch exists and \code{FALSE} otherwise.
}
\details{
\url{https://developer.github.com/v3/git/refs/#get-a-reference}
}
