% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branches-deprecated.R
\name{branches_exist}
\alias{branches_exist}
\title{Determine whether branches exist in the specified repository.}
\usage{
branches_exist(
  branches,
  repo,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{branches}{(character) The name of the branch.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A logical vector containing \code{TRUE} or \code{FALSE} for each branch specified.
}
\description{
This function returns \code{TRUE} if the branch exists and \code{FALSE} otherwise.
}
\details{
\url{https://developer.github.com/v3/git/refs/#get-a-reference}
}
