% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-commits.R
\name{compare_commits}
\alias{compare_commits}
\title{Compare two commits}
\usage{
compare_commits(repo, base, head, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified as "owner/repo".}

\item{base}{(string) The reference of the base commit.}

\item{head}{(string) The reference of the commit to compare.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the commits, with the following columns
(see \href{https://developer.github.com/v3/repos/commits/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{sha}: The SHA of the commit.
\item \strong{message}: The commit message recorded in GitHub.
\item \strong{author_name}: The author of the commit.
\item \strong{author_email}: The author's email address.
\item \strong{committer_name}: The committer of the commit.
\item \strong{committer_email}: The committer's email address.
\item \strong{date}: The date and time of the commit.
\item \strong{url}: The URL to get the commit details from GitHub.
\item \strong{tree_sha}: The SHA of the git tree.
\item \strong{tree_url}: The URL to get the tree details from GitHub.
\item \strong{parent_sha}: The SHA(s) of the parent commit(s).
\item \strong{parent_url}: The URL(s) to get the parent's details from GitHub.
}
}
\description{
This function returns information about all the commits made between the two specified
commits. The commits can be specifed by any git reference, i.e. branch, tag or SHA. Note:
This function will only return a maximum of 250 commits. If more are required, use
\code{\link[=view_commits]{view_commits()}}.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
