% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{compare_files}
\alias{compare_files}
\title{View file changes made between two commits}
\usage{
compare_files(base, head, repo, ...)
}
\arguments{
\item{base}{(string) Either a SHA, branch or tag used to identify the base commit.}

\item{head}{(string) Either a SHA, branch or tag used to identify the head commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{compare_files()} returns a tibble of file properties.

\strong{File Properties:}
\itemize{
\item \strong{path}: The path to the file within the repository.
\item \strong{sha}: The SHA of the file blob.
\item \strong{status}: Whether the file was "added", "modified" or "deleted".
\item \strong{additions}: The number of lines added in the file.
\item \strong{deletions}: The number of lines deleted in the file.
\item \strong{changes}: The number of lines changed in the file.
\item \strong{patch}: The git patch for the file.
\item \strong{html_url}: The URL of the blob's web page in GitHub.
}
}
\description{
\code{compare_files()} summarises the file changes made between two commits in a table with the
properties as columns and a row for each file. The \code{base} commit must be in the history
of the \code{head} commit.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
}
\examples{
\dontrun{

  # View the files changes made between the current main branch and a release
  compare_files("main", "0.8.7", "ChadGoymer/githapi")

}

}
