% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-blobs.R
\name{create_blobs}
\alias{create_blobs}
\title{Create blobs (files).}
\usage{
create_blobs(repo, contents, encoding = "utf-8",
  token = getOption("github.token"), api = getOption("github.api"),
  ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{contents}{(character) The contents of the files.}

\item{encoding}{(character, optional) The encoding used for the content. Currently,
\code{"utf-8"} and \code{"base64"} are supported. Can supply either a single string, which will apply
to every blob, or a character vector the same length as the contents, specifying the
encoding for each blob separately. Default: \code{"utf-8"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the blobs, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{sha}: The SHA of the blob.
\item \strong{url}: The URL to get the blob details from GitHub.
}
}
\description{
This function creates the specified blobs within a repository in GitHub.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#create-a-blob}
}
