% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{create_branch}
\alias{create_branch}
\title{Create a branch in a repository}
\usage{
create_branch(name, ref, repo, ...)
}
\arguments{
\item{name}{(string) The name of the branch.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_branch()} returns a list of the branch's properties.

\strong{Branch Properties:}
\itemize{
\item \strong{name}: The name of the branch.
\item \strong{ref}: The full Git reference of the branch.
\item \strong{sha}: The commit SHA the branch is pointing at.
}
}
\description{
This function creates a new branch in the specified repository in GitHub. It must be
pointed at a commit by providing a Git reference, which can be either a SHA, branch or
tag. For a branch, the head commit is used.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#create-a-reference}
}
}
\examples{
\dontrun{

  create_branch(
    name = "new-branch",
    ref  = "master",
    repo = "ChadGoymer/githapi")

}

}
