% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{create_card}
\alias{create_card}
\title{Create a card in a GitHub project}
\usage{
create_card(
  content_id,
  content_type,
  note,
  column,
  project,
  repo,
  user,
  org,
  ...
)
}
\arguments{
\item{content_id}{(integer, optional) The issue or pull request id associated with this
card. Note: \code{content_type} is also required.}

\item{content_type}{(string, optional) Whether the \code{content_id} identifies an \code{"Issue"}
or a \code{"PullRequest"}.}

\item{note}{(string, optional) The card's note content. Not used if \code{content_id} is
specified.}

\item{column}{(integer or string) Either the ID or name of the column.}

\item{project}{(integer or string) Either the project number or name.}

\item{repo}{(string, optional) The repository specified in the format: \code{owner/repo}.}

\item{user}{(string, optional) The login of the user.}

\item{org}{(string, optional) The name of the organization.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_card()} returns a list of the card properties.

\strong{Card Properties:}
\itemize{
\item \strong{id}: The ID of the card.
\item \strong{content_id}: The ID of the issue or pull request.
\item \strong{note}: The content of a note.
\item \strong{archived}: Whether the card has been archived.
\item \strong{creator}: The creator of the note.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
This function creates a new card in the specified column in a project in GitHub. The
card can either contain an existing issue or pull request or you can create a note.
}
\details{
You can create a card in a project associated with either a repository, user or
organization, by supplying them as an input, as long as you have appropriate permissions.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/projects/cards/#create-a-project-card}
}
}
\examples{
\dontrun{
  # Create an issue card in a repository project
  create_card(
    content_id   = 1,
    content_type = "Issue",
    column       = "Test column",
    project      = "Test project",
    repo         = "ChadGoymer/test-githapi")

  # Create a pull request card in a user's project
  create_card(
    content_id   = 2,
    content_type = "PullRequest",
    column       = "Test column",
    project      = "Test project",
    user         = "ChadGoymer")

  # Create a note card in an organization's project
  create_card(
    note    = "This is a note",
    column  = "Test column",
    project = "Test project",
    org     = "HairyCoos")
}

}
