% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-commits.R
\name{create_commit}
\alias{create_commit}
\title{Create a commit}
\usage{
create_commit(message, tree, parents, committer, author, repo,
  token = getOption("github.token"), api = getOption("github.api"),
  ...)
}
\arguments{
\item{message}{(string) The commit message}

\item{tree}{(string) The SHA of the tree object this commit points to}

\item{parents}{(character, optional) The SHAs of the commits that were the parents of this
commit. If not specified the commit will be written as a root commit. For a single parent,
one SHA should be provided; for a merge commit, a character vector of more than one should be
provided.}

\item{committer}{(list, optional) The name and email address of the committer. This needs
to be specified as a named list, e.g. \code{list(name = "Bob Smith", email = "bob.smith@acme.com")}.
If not specified then the authenticated user is used.}

\item{author}{(list, optional) The name and email address of the author. This needs to be
specified as a named list, e.g. \code{list(name = "Bob Smith", email = "bob.smith@acme.com")}.
If not specified then the authenticated user is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the commit, with the following columns
(see \href{https://developer.github.com/v3/repos/commits/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{sha}: The SHA of the commit.
\item \strong{message}: The commit message recorded in GitHub.
\item \strong{author_name}: The author of the commit.
\item \strong{author_email}: The author's email address.
\item \strong{committer_name}: The committer of the commit.
\item \strong{committer_email}: The committer's email address.
\item \strong{date}: The date and time of the commit.
\item \strong{url}: The URL to get the commit details from GitHub.
\item \strong{tree_sha}: The SHA of the git tree.
\item \strong{tree_url}: The URL to get the tree details from GitHub.
\item \strong{parent_sha}: The SHA(s) of the parent commit(s).
\item \strong{parent_url}: The URL(s) to get the parent's details from GitHub.
}
}
\description{
This function creates a commit object in a GitHub repository. The commit must point to a
tree object that has already been uploaded and identified by it's SHA. In addition at
least one parent commit must be specified.
}
\details{
\url{https://developer.github.com/v3/git/commits/#create-a-commit}
}
