% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{create_file}
\alias{create_file}
\alias{update_file}
\alias{delete_file}
\title{Create, update or delete a file in a new commit}
\usage{
create_file(content, path, branch, message, repo, author, committer, ...)

update_file(content, path, branch, message, repo, author, committer, ...)

delete_file(path, branch, message, repo, author, committer, ...)
}
\arguments{
\item{content}{(string) The content of the file specified as a single string.}

\item{path}{(string) The path to create the file at, within the repository.}

\item{branch}{(string) The name of the branch to make the new commit on.}

\item{message}{(string) The commit message.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{author}{(list, optional) A the name and email address of the user who wrote the
changes in the commit.}

\item{committer}{(list, optional) A the name and email address of the user who created
the commit.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_file()}, \code{update_file()} and \code{delete_file()} return a list of the commit
properties.

\strong{Commit Properties:}
\itemize{
\item \strong{sha}: The commit SHA.
\item \strong{message}: The commit message.
\item \strong{author_name}: The author's name.
\item \strong{author_email}: The author's email address.
\item \strong{committer_name}: The committer's name.
\item \strong{committer_email}: The committer's email address.
\item \strong{tree_sha}: The SHA of the file tree.
\item \strong{parent_sha}: The commit SHA of the parent(s).
\item \strong{date}: The date the commit was made.
}
}
\description{
These functions create, update or delete a file in a repository in GitHub by creating a
new commit on the specified branch. If the file already exists \code{\link[=create_file]{create_file()}} throws an
error and if the file does not exist \code{\link[=delete_file]{delete_file()}} throws as error.
}
\details{
The \code{author} and \code{committer} arguments are optional and if not supplied the current
authenticated user is used. However, if you want to set them explicitly you must specify
a named list with \code{name} and \code{email} as the elements (see examples).

Note: The GitHub API imposes a file size limit of 1MB for this request. For larger files
use the \code{\link[=upload_files]{upload_files()}} function.
}
\examples{
\dontrun{

  # Create a new file on the master branch
  create_file(
    content = "# This is a new file\\\\n\\\\n Created by `create_files()`",
    path    = "new-file.md",
    branch  = "master",
    message = "Created a new file with create_file()",
    repo    = "ChadGoymer/githapi")

  # Create a new file on the master branch specifying an author and committer
  create_file(
    content   = "# This is a new file\\\\n\\\\n Created by `create_files()`",
    path      = "new-file.md",
    branch    = "master",
    message   = "Created a new file with create_file()",
    repo      = "ChadGoymer/githapi",
    author    = list(name = "Bob",  email = "bob@acme.com"),
    committer = list(name = "Jane", email = "jane@acme.com"))

  # Update an existing file on the master branch
  update_file(
    content = "# This is an updated file\\\\n\\\\n Updated by `update_files()`",
    path    = "new-file.md",
    branch  = "master",
    message = "Updated an existing file with update_file()",
    repo    = "ChadGoymer/githapi")

  # Delete an existing file on the master branch
  delete_file(
    path    = "new-file.md",
    branch  = "master",
    message = "Deleted an existing file with delete_file()",
    repo    = "ChadGoymer/githapi")
}

}
