% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{create_gist}
\alias{create_gist}
\title{Create a gist in github}
\usage{
create_gist(files, description, public = FALSE, ...)
}
\arguments{
\item{files}{(list) The file contents with the list names as the filenames.}

\item{description}{(string, optional) A description for the gist.}

\item{public}{(boolean, optional) Whether the gist is public. Default: \code{FALSE}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_gist()} returns a list of the gist's properties.

\strong{Gist Properties:}
\itemize{
\item \strong{id}: The id for the gist
\item \strong{description}: The description of the gist
\item \strong{files}: A tibble of file properties:
\itemize{
\item \strong{filename}: The name of the file
\item \strong{type}: The type of file
\item \strong{content}: The file content
\item \strong{size}: The size of the file in bytes
\item \strong{truncated}: Whether the file content has been truncated
}
\item \strong{owner}: The login if the gist's owner
\item \strong{public}: Whether the gist is public
\item \strong{html_url}: The address of the gist's web page.
\item \strong{created_at}: The time and date the gist was created.
\item \strong{updated_at}: The time and date the gist was last updated.
}
}
\description{
This function creates a new gist in GitHub. You can specify one or more files by providing
a named list (see examples).
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/gists/#create-a-gist}
}
}
\examples{
\dontrun{

  # Create a gist with a single file
  create_gist(files = list(hello_world.R = "print(\"Hello World!\")"))

  # Create a gist with multiple files
  create_gist(
    files = list(
      hello_world.R    = "print(\"Hello World!\")",
      `hello_world-fn.R` = "helloworld <- function() print(\"Hello World!\")"),
    description = "A new gist")

  # Create a public gist
  create_gist(
    files       = list(hello_world.R = "print(\"Hello World!\")"),
    description = "A new gist",
    public      = TRUE)

}

}
