% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{create_issue}
\alias{create_issue}
\title{Create an issue in a repository}
\usage{
create_issue(title, repo, body, assignees, labels, milestone, ...)
}
\arguments{
\item{title}{(string) The title of the issue.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{body}{(string, optional) The contents of the issue.}

\item{assignees}{(character, optional) Logins for Users to assign to this
issue. NOTE: Only users with push access can set assignees for new issues.}

\item{labels}{(character, optional) Labels to associate with this issue.
NOTE: Only users with push access can set labels for new issues.}

\item{milestone}{(character or integer, optional) The title or number of the
milestone to associate this issue with. NOTE: Only users with push access
can set the milestone for new issues.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_issue()} returns a list of the issue's properties.

\strong{Issue Properties:}
\itemize{
\item \strong{number}: The number assigned to the issue.
\item \strong{title}: The title of the issue.
\item \strong{body}: The body contents of the issue.
\item \strong{assignees}: The users assigned to the issue.
\item \strong{labels}: The labels attached to the issue.
\item \strong{milestone}: The milestone assigned to the issue.
\item \strong{state}: The state of the issue - either \code{"open"} or \code{"closed"}.
\item \strong{repository}: The repository the issue is in.
\item \strong{pull_request}: Whether the issue is a pull request.
\item \strong{html_url}: The URL of the issue's web page in GitHub.
\item \strong{creator}: The creator's login.
\item \strong{created_at}: When the issue was created.
\item \strong{updated_at}: When the issue was last updated.
\item \strong{closed_at}: When the issue was closed.
}
}
\description{
This function creates a new issue for the specified repository in GitHub. It
can also be used to assign the issue to a user and add labels or a milestone.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/issues#create-an-issue}
}
}
\examples{
\dontrun{

  create_issue(
    title     = "user issue",
    repo      = "ChadGoymer/test-issues",
    body      = "This is an issue to test create_issue()",
    assignees = "ChadGoymer",
    labels    = "feature",
    milestone = "release-1.0"
  )

  create_issue(
    title     = "organization issue",
    repo      = "HairyCoos/test-issues",
    body      = "This is an issue to test create_issue()",
    assignees = "ChadGoymer"
  )

}

}
