% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{create_label}
\alias{create_label}
\title{Create an issue label in a repository}
\usage{
create_label(name, repo, description, color = random_color(), ...)
}
\arguments{
\item{name}{(string) The name of the label.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{description}{(string, optional) A description of the label.}

\item{color}{(string, optional) Either the color name (see \code{\link[grDevices:colors]{grDevices::colors()}}) or a
hexidecimal color code (see \href{http://www.color-hex.com/}{color-hex.com}). If not
supplied a color is chosen at random.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_label()} returns a list of the label's properties.

\strong{Label Properties:}
\itemize{
\item \strong{name}: The name of the label.
\item \strong{color}: The color of the label in hexidecimal code.
\item \strong{description}: The description of the label.
}
}
\description{
This function creates a new label for issues and pull requests for the specified
repository in GitHub. It can also be used to set the colour and description.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/labels/#create-a-label}
}
}
\examples{
\dontrun{

  create_label(
    name        = "new-label",
    repo        = "ChadGoymer/githapi",
    color       = "blue",
    description = "This is a detailed label")

}

}
