% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{create_milestone}
\alias{create_milestone}
\title{Create a milestone in a repository}
\usage{
create_milestone(title, repo, description, due_on, state = "open", ...)
}
\arguments{
\item{title}{(string) The title of the milestone.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{description}{(string, optional) A description of the milestone.}

\item{due_on}{(string, optional) The milestone due date. This is in the format:
\code{YYYY-MM-DD}.}

\item{state}{(string, optional) The state of the milestone. Either \code{"open"} or \code{"closed"}.
Default: \code{"open"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_milestone()} returns a list of the milestone's properties.

\strong{Milestone Properties:}
\itemize{
\item \strong{number}: The number assigned to the milestone within the repository.
\item \strong{title}: The title of the milestone.
\item \strong{description}: The description of the milestone.
\item \strong{state}: The state of the milestone - either "open" or "closed".
\item \strong{open_issues}: The number of open issues within the milestone.
\item \strong{closed_issues}: The number of closed issues within the milestone.
\item \strong{html_url}: The URL of the milestone's web page in GitHub.
\item \strong{creator}: The creator's login.
\item \strong{created_at}: When the milestone was created.
\item \strong{updated_at}: When the milestone was last updated.
\item \strong{due_on}: When the milestone is due.
\item \strong{closed_at}: When the milestone was closed.
}
}
\description{
This function creates a new milestone for the specified repository in GitHub. It can also be
used to set a due date.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/milestones/#create-a-milestone}
}
}
\examples{
\dontrun{

  create_milestone(
    title       = "test milestone",
    repo        = "ChadGoymer/test-githapi",
    description = "This is a test milestone",
    due_on      = "2030-01-01 00:00:00")

}

}
