% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{create_project}
\alias{create_project}
\title{Create a GitHub project}
\usage{
create_project(name, body, repo, org, ...)
}
\arguments{
\item{name}{(string) The name of the project.}

\item{body}{(string) The description of the project.}

\item{repo}{(string, optional) The repository specified in the format: \code{owner/repo}.}

\item{org}{(string, optional) The name of the organization.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_project()} returns a list of the project properties.

\strong{Project Properties:}
\itemize{
\item \strong{id}: The ID of the project.
\item \strong{number}: The number of the project for the repository, user or organisation.
\item \strong{name}: The name given to the project.
\item \strong{body}: The description given to the project.
\item \strong{state}: Whether the project is "open" or "closed".
\item \strong{creator}: The user who created the project.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
\item \strong{html_url}: The URL to view the project.
}
}
\description{
This function creates a new project in GitHub. The project will be empty so you will need
to add columns and cards separately.
}
\details{
You can create a project associated with either a repository or organisation, by
supplying them as an input, as long as you have appropriate permissions. If no repository
or organisation is specified the project is created for the authenticated user.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/projects/#create-a-repository-project}
\item \url{https://developer.github.com/v3/projects/#create-an-organization-project}
\item \url{https://developer.github.com/v3/projects/#create-a-user-project}
}
}
\examples{
\dontrun{
  create_project(
    name = "Repo project",
    body = "This is a repository's project",
    repo = "ChadGoymer/test-githapi")

  create_project(
    name = "User project",
    body = "This is a user's project")
}

}
