% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{create_pull_request}
\alias{create_pull_request}
\title{Create an pull request in a repository}
\usage{
create_pull_request(
  title,
  repo,
  head,
  base,
  body,
  assignees,
  reviewers,
  labels,
  milestone,
  ...
)
}
\arguments{
\item{title}{(string) The title of the pull request.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{head}{(string) The name of the branch where your changes are implemented. For
cross-repository pull requests, prefix \code{head} with an owner, e.g. \code{"username:branch"}.}

\item{base}{(string) The name of the branch you want the changes pulled into. This should
be an existing branch in the specified repository.}

\item{body}{(string, optional) The contents of the pull request.}

\item{assignees}{(character, optional) Logins for Users to assign to this pull request.
NOTE: Only users with push access can set assignees for new pull requests.}

\item{reviewers}{(character, optional) Logins for Users to review this pull request.
NOTE: Only users with push access can set reviewers for new pull requests.}

\item{labels}{(character, optional) Labels to associate with this pull request. NOTE: Only
users with push access can set labels for new pull requests.}

\item{milestone}{(character or integer, optional) The title or number of the milestone to
associate this pull request with. NOTE: Only users with push access can set the
milestone for new pull requests.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_pull_request()} returns a list of the pull request's properties.

\strong{Pull Request Properties:}
\itemize{
\item \strong{number}: The number assigned to the pull request.
\item \strong{title}: The title of the pull request.
\item \strong{body}: The body contents of the pull request.
\item \strong{head_sha}: The SHA of the commit to merge in.
\item \strong{head_ref}: The reference, or branch, to merge in.
\item \strong{head_repo}: The repository containing the branch to merge in.
\item \strong{base_sha}: The SHA of the commit to merge onto.
\item \strong{base_ref}: The reference, or branch, to merge onto.
\item \strong{merge_sha}: The SHA of the merge commit, if the merge has been completed.
\item \strong{assignees}: The users assigned to the pull request.
\item \strong{reviewers}: The users reviewing the pull request.
\item \strong{labels}: The labels attached to the pull request.
\item \strong{milestone}: The milestone assigned to the pull request.
\item \strong{state}: The state of the pull request - either \code{"open"} or \code{"closed"}.
\item \strong{repository}: The repository the pull request is in.
\item \strong{html_url}: The URL of the pull request's web page in GitHub.
\item \strong{diff_url}: The URL of the pull request's diff web page in GitHub.
\item \strong{creator}: The creator's login.
\item \strong{created_at}: When the pull request was created.
\item \strong{updated_at}: When the pull request was last updated.
\item \strong{mergeable}: Whether the pull request can be merged.
\item \strong{rebaseable}: Whether the pull request can be rebased.
\item \strong{merged}: Whether the pull request has been merged
\item \strong{merged_by}: Who merged the pull request.
\item \strong{merged_at}: When the pull request was merged.
\item \strong{closed_at}: When the pull request was closed.
}
}
\description{
This function creates a new pull request for the specified repository in GitHub. It can
also be used to assign the pull request to a user, request reviewers and add labels or a
milestone.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/pulls/#create-a-pull-request}
\item \url{https://developer.github.com/v3/issues/#edit-an-issue}
\item \url{https://developer.github.com/v3/pulls/review_requests/#create-a-review-request}
}
}
\examples{
\dontrun{

  create_pull_request(
    title = "test pull request",
    repo  = "ChadGoymer/test-githapi",
    head  = "test-pulls",
    base  = "master",
    body  = "This is a pull request to test create_pull_request()")

  create_pull_request(
    title     = "test assigned pull request",
    repo      = "ChadGoymer/test-githapi",
    head      = "test-pulls-2",
    base      = "master",
    body      = "This is a pull request to test create_pull_request()",
    assignees = "ChadGoymer",
    reviewers = c("BobSmith", "JaneJones"),
    labels    = "enhancement",
    milestone = "Release-1.0")

}

}
