% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{create_release}
\alias{create_release}
\title{Create a release in a repository}
\usage{
create_release(
  tag,
  repo,
  name,
  body,
  ref,
  draft = FALSE,
  prerelease = FALSE,
  ...
)
}
\arguments{
\item{tag}{(string) The name of the tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{name}{(string, optional) The name of the release.}

\item{body}{(string, optional) The description of the release.}

\item{ref}{(string, optional) Either a SHA or branch used to identify the commit. If the
\code{tag} exists this \code{ref} is ignored. If not supplied the repositories default branch is
used.}

\item{draft}{(boolean, optional) Whether the release is a draft. Default: \code{FALSE}.}

\item{prerelease}{(boolean, optional) Whether the release is a pre-release.
Default: \code{FALSE}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_release()} returns a list of the release's properties.

\strong{Release Properties:}
\itemize{
\item \strong{id}: The id of the release.
\item \strong{tag}: The tags associated with the release.
\item \strong{name}: The name of the release.
\item \strong{body}: The description of the release.
\item \strong{commit}: The commit associated with the release.
\item \strong{draft}: Whether the release is draft.
\item \strong{prerelease}:  Whether it is a pre-release.
\item \strong{author_login}: The author's account login.
\item \strong{assets}: The name of the assets associated with the release.
\item \strong{html_url}: The address of the release's web page.
\item \strong{created_at}: The time and date the release was created.
\item \strong{published_at}: The time and date the release was published.
}
}
\description{
This function creates a new release in the specified repository in GitHub. It must be
pointed at a commit by providing an existing tag or a Git reference, which can be either a
SHA or branch. If the tag does not exist it is created pointing at the commit; if it does
exist then the reference is ignored.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/releases/#create-a-release}
}
}
\examples{
\dontrun{

  # Create a release on the default branch
  create_release(
    tag  = "1.0.0",
    repo = "ChadGoymer/githapi",
    name = "Initial production release",
    body = "This is a release created by create_release()")

  # Create a draft pre-release
  create_release(
    tag        = "1.0.9000",
    repo       = "ChadGoymer/githapi",
    name       = "Draft 1.1 release",
    body       = "This is a release created by create_release()",
    ref        = "dev-branch",
    draft      = TRUE,
    prerelease = TRUE)

}

}
