% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-releases-deprecated.R
\name{create_releases}
\alias{create_releases}
\title{Create releases at specified commits.}
\usage{
create_releases(
  tags,
  commits,
  names,
  bodies,
  repo,
  draft = FALSE,
  prerelease = FALSE,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{tags}{(character) The tag names.}

\item{commits}{(character) The commits to tag. Can be either a SHA or a branch name.}

\item{names}{(character) The names of the releases.}

\item{bodies}{(character) The bodies for the releases.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{draft}{(logical, optional) Whether the releases are draft. Default: \code{FALSE}.}

\item{prerelease}{(logical, optional) Whether the releases are a prerelease. Default: \code{FALSE}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the releases, with the following columns
(see \href{https://developer.github.com/v3/repos/releases/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{id}: The unique ID assigned by GitHub.
\item \strong{tag_name}: The name of the tag associated with the release.
\item \strong{name}: The name, or title, of the release.
\item \strong{body}: The body, or description, of the release.
\item \strong{author_login}: The author of the release.
\item \strong{draft}: Whether the release is a draft.
\item \strong{prerelease}: Whether the release is a pre-release.
\item \strong{target_commitish}: The commit or branch the release was applied to.
\item \strong{created_at}: The date and time of creation.
\item \strong{published_at}: The date and time of publication.
\item \strong{assets}: The names of any assets associated with the release.
\item \strong{zipball_url}: The URL to download a zip file of the release commit.
\item \strong{url}: The URL to view the release details on GitHub.
}
}
\description{
This function creates releases for specified commits. Tag names must be specified for each
commit along with release names and bodies (descriptions). You can also specify whether the
releases are draft and whether they are pre-releases.
}
\details{
\url{https://developer.github.com/v3/repos/releases/#create-a-release}
}
