% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{create_status}
\alias{create_status}
\title{Create a status for a commit}
\usage{
create_status(
  state,
  ref,
  repo,
  description,
  target_url,
  context = "default",
  ...
)
}
\arguments{
\item{state}{(string) The state of the status. Can be one of \code{"error"},
\code{"failure"}, \code{"pending"}, or \code{"success"}.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{description}{(string, optional) A short description of the status.}

\item{target_url}{(string, optional) The target URL to associate with this
status. This URL will be linked from the GitHub UI to allow users to easily
see the source of the status.}

\item{context}{(string, optional) A string label to differentiate this status
from the status of other systems. Default: \code{"default"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_status()} returns a list of the status's properties.

\strong{Status Properties:}
\itemize{
\item \strong{id}: The id of the release.
\item \strong{state}: The state of the status.
\item \strong{description}: The description of the status.
\item \strong{target_url}: The URL linked to the status.
\item \strong{context}: The context of the status.
\item \strong{creator}: The creator of the status.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
This function creates a new status for a commit in the specified repository
in GitHub. The status can also have a \code{description} and a target URL,
allowing a user to link to more information. The \code{context} allows you to
differentiate the status from others. The status can then be updated, from
\code{pending} to \code{success} say, by using the same \code{context}.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#create-a-commit-status}
}
}
\examples{
\dontrun{

  # Create a status on the main branch
  create_status(
    state       = "pending",
    ref         = "main",
    repo        = "ChadGoymer/test-githapi",
    description = "This is a pending status",
    target_url  = "https://goymer.me/githapi",
    context     = "test/githapi"
  )

  # Update the status on the main branch
  create_status(
    state       = "success",
    ref         = "main",
    repo        = "ChadGoymer/test-githapi",
    description = "This is a success status",
    target_url  = "https://goymer.me/githapi",
    context     = "test/githapi"
  )

}

}
