% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{create_tag}
\alias{create_tag}
\title{Create a tag in a repository}
\usage{
create_tag(name, repo, ref = "master", ...)
}
\arguments{
\item{name}{(string) The name of the tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) Either a SHA, branch or tag used to identify the commit
the tag is pointing at. Default: \code{"master"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_tag()} returns a list of the tag's properties.

\strong{Tag Properties:}
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{ref}: The full Git reference of the tag.
\item \strong{sha}: The commit SHA the tag is pointing at.
}
}
\description{
This function creates a new tag in the specified repository in GitHub. It must be
pointed at a commit by providing a Git reference, which can be either a SHA, branch or
tag. For a branch, the head commit is used.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#create-a-reference}
}
}
\examples{
\dontrun{
  create_tag(
    name = "new-tag",
    repo = "ChadGoymer/test-githapi",
    ref  = "master")
}

}
