% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{create_team}
\alias{create_team}
\title{Create a team in a GitHub organization}
\usage{
create_team(
  name,
  org,
  description,
  maintainers,
  repo_names,
  privacy,
  parent_team,
  ...
)
}
\arguments{
\item{name}{(string) The name of the team.}

\item{org}{(string) The login of the organization.}

\item{description}{(string, optional) The description of the team.}

\item{maintainers}{(character, optional) The logins of organization members
to add as maintainers of the team. If you do not specify any maintainers,
then you will automatically become a team maintainer when you create a new
team.}

\item{repo_names}{(character, optional) The full name (e.g.
"organization-name/repository-name") of repositories to add the team to.}

\item{privacy}{(string, optional) The level of privacy this team should have.
The options are:

For a non-nested team:
\itemize{
\item \code{"secret"}: only visible to organization owners and members of this team.
\item \code{"closed"}: visible to all members of this organization.
}

Default: \code{"secret"}.

For a parent or child team:
\itemize{
\item \code{"closed"}: visible to all members of this organization.
}

Default for child team: \code{"closed"}}

\item{parent_team}{(integer or string, optional) The ID or name of a team to
set as the parent team.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{create_team()} returns a list of the team's properties.

\strong{Team Properties:}
\itemize{
\item \strong{id}: The ID of the team.
\item \strong{name}: The name of the team.
\item \strong{organization}: The organization the team is associated with.
\item \strong{slug}: The team slug name.
\item \strong{description}: The description of the team.
\item \strong{privacy}: The privacy setting of the team - either "closed" or "secret".
\item \strong{permission}: The default repository permissions of the team.
\item \strong{parent}: The parent team.
\item \strong{members_count}: The number of members.
\item \strong{repos_count}: The number of repositories the team has access to.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
\item \strong{html_url}: The URL of the team page in GitHub
}
}
\description{
This function creates a new team in the specified organization in GitHub. It
can also be used to specify the maintainers, the repositories to add the team
to and a parent team.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/teams#create-a-team}
}
}
\examples{
\dontrun{

  # Create a team in an organization
  create_team("TestTeam", "HairyCoos")

  # Create a team and specify maintainers
  create_team(
    name        = "TestTeam2",
    org         = "HairyCoos",
    maintainers = c("ChadGoymer", "chris-walker1")
  )

  # Create a team and set the privacy
  create_team("TestTeam3", "HairyCoos", privacy = "closed")

  # Create a team and specify a parent team
  create_team("TestTeam4", "HairyCoos", parent_team = "TestTeam3")

}

}
