% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-trees.R
\name{create_tree}
\alias{create_tree}
\title{Create a tree.}
\usage{
create_tree(repo, paths, modes, types, shas = NA, contents = NA,
  base_tree = NA, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{paths}{(character) The paths to the files or directories.}

\item{modes}{(character) The file modes; one of \code{"100644"} for file (blob), \code{"100755"} for
executable (blob), \code{"040000"} for subdirectory (tree), \code{"160000"} for submodule (commit), or
\code{"120000"} for a blob that specifies the path of a symlink.}

\item{types}{(character) The object types. Either \code{"blob"}, \code{"tree"}, or \code{"commit"}.}

\item{shas}{(character) The SHAs of the objects in the tree.}

\item{contents}{(character) The contents of each file. GitHub will write this blob out and use
that SHA for this entry. Use either this, or sha.}

\item{base_tree}{(string, optional) The SHAs of the tree you want to update with new data.
Default: \code{NA}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the trees, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{sha}: The SHA of the tree.
\item \strong{url}: The URL to get the tree details from GitHub.
}
}
\description{
This function creates the specified tree within a repository in GitHub. The tree is either
specified by supplying SHAs of existing blobs and trees, or by specifying the text contents
of files, which are uploaded as blobs.
}
\details{
\url{https://developer.github.com/v3/git/trees/#create-a-tree}
}
