% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{delete_branch}
\alias{delete_branch}
\title{Delete a branch from a repository}
\usage{
delete_branch(branch, repo, ...)
}
\arguments{
\item{branch}{(string) The name of the branch.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_branch()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a branch from a repository, as long as you have appropriate
permissions. Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#delete-a-reference}
}
}
\examples{
\dontrun{
  delete_branch("new-branch", repo = "ChadGoymer/test-githapi")
}

}
