% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{delete_card}
\alias{delete_card}
\title{Delete a card in a GitHub project}
\usage{
delete_card(card, ...)
}
\arguments{
\item{card}{(integer) The ID of the card.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_card()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a card in a GitHub project. Care should be taken as it
will not be recoverable. Instead, you may way to archive the card with
\code{\link[=update_card]{update_card()}}.
}
\details{
You can delete a card associated with either a repository, user or
organization, by supplying them as an input, as long as you have appropriate
permissions.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/projects#delete-a-project-card}
}
}
\examples{
\dontrun{

  # Delete a card
  delete_card(card = 123456)

}

}
