% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-contents.R
\name{delete_files}
\alias{delete_files}
\title{Delete files.}
\usage{
delete_files(paths, messages, branches, committer, author, repo,
  token = getOption("github.token"), api = getOption("github.api"),
  ...)
}
\arguments{
\item{paths}{(character) The paths to the files in the repository.}

\item{messages}{(character) The commit messages to use. If a single message is provided it
is used for all commits with the file added as a suffix.}

\item{branches}{(character, optional) The branches to make the commits on. If a single branch
is provided all commits are made on it. If not specified the default branch for the repository
is used.}

\item{committer}{(list, optional) The name and email address of the committer. This needs
to be specified as a named list, e.g. \code{list(name = "Bob Smith", email = "bob.smith@acme.com")}.
If not specified then the authenticated user is used.}

\item{author}{(list, optional) The name and email address of the author. This needs to be
specified as a named list, e.g. \code{list(name = "Bob Smith", email = "bob.smith@acme.com")}.
If not specified then the authenticated user is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the files, with the following columns
(see \href{https://developer.github.com/v3/repos/contents/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{commit_message}: The commit message recorded in GitHub.
\item \strong{commit_sha}: The SHA of the commit.
\item \strong{commit_url}: The URL to the commit.
\item \strong{commit_author}: The author of the commit.
\item \strong{commit_committer}: The committer of the commit.
\item \strong{commit_tree_sha}: The SHA for the git tree.
\item \strong{commit_tree_url}: The URL for the git tree.
\item \strong{commit_parent_sha}: The SHA of the parent commit.
\item \strong{commit_parent_url}: The URL to the parent commit.
}
}
\description{
This function deletes the specified files from a repository in GitHub. It adds a new commit
for each file deleted with the specifed commit message and on the specified branch. The
committer and author can be set explicitly if necessary.
}
\details{
\url{https://developer.github.com/v3/repos/contents/#delete-a-file}
}
