% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{delete_gist}
\alias{delete_gist}
\title{Delete a gist in GitHub}
\usage{
delete_gist(gist, ...)
}
\arguments{
\item{gist}{(string) The id of the gist.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_gist()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a gist from github, as long as you have appropriate permissions.
Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/gists/#delete-a-gist}
}
}
\examples{
\dontrun{

  delete_gist("806dca6b09a39e7b6326a0c8137583e6")

}

}
