% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{delete_label}
\alias{delete_label}
\title{Delete a label from a repository}
\usage{
delete_label(label, repo, ...)
}
\arguments{
\item{label}{(string) The name of the label.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_label()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a label from a repository, as long as you have appropriate
permissions. Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/labels/#delete-a-label}
}
}
\examples{
\dontrun{

  delete_label("new-label", repo = "ChadGoymer/githapi")

}

}
