% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{delete_milestone}
\alias{delete_milestone}
\title{Delete a milestone from a repository}
\usage{
delete_milestone(milestone, repo, ...)
}
\arguments{
\item{milestone}{(string or character) The number or title of the milestone.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_milestone()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a milestone from a repository, as long as you have appropriate
permissions. Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/milestones/#delete-a-milestone}
}
}
\examples{
\dontrun{
  # Delete a milestone
  delete_milestone(
    milestone = "test milestone",
    repo      = "ChadGoymer/test-githapi")
}

}
