% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{delete_release}
\alias{delete_release}
\title{Delete a release from a repository}
\usage{
delete_release(release, repo, ...)
}
\arguments{
\item{release}{(string) The id or tag of the release.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_release()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a release from a repository, as long as you have
appropriate permissions. Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#delete-a-release}
}
}
\examples{
\dontrun{

  delete_release("1.0.0", repo = "ChadGoymer/githapi")

}

}
