% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{delete_releases}
\alias{delete_releases}
\title{Delete releases.}
\usage{
delete_releases(repo, tags, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{tags}{(character) The tag names associated with the releases.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A named list containing \code{TRUE} if the release was deleted. An error is thrown otherwise.
}
\description{
This function deletes the specified releases. It also deletes the tags associated with the
releases from the repository.
}
\details{
\url{https://developer.github.com/v3/repos/releases/#delete-a-release}
}
