% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{delete_repository}
\alias{delete_repository}
\title{Delete a user or organization repository}
\usage{
delete_repository(repo, ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_repository()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a repository from GitHub, as long as you have appropriate permissions.
Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/#delete-a-repository}
}
}
\examples{
\dontrun{
  # Delete a user's repository
  delete_repository("ChadGoymer/user-repository")

  # Delete a organization's repository
  delete_repository("HairyCoos/org-repository")
}

}
