% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{delete_tag}
\alias{delete_tag}
\title{Delete a tag from a repository}
\usage{
delete_tag(tag, repo, ...)
}
\arguments{
\item{tag}{(string) The name of the tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_tag()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a tag from a repository, as long as you have appropriate permissions.
Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#delete-a-reference}
}
}
\examples{
\dontrun{
  delete_tag("new-tag", repo = "ChadGoymer/test-githapi")
}

}
