% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{delete_tags}
\alias{delete_tags}
\title{Delete tags.}
\usage{
delete_tags(repo, tags, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{tags}{(character) The tag names.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A named list containing \code{TRUE} if the tag was deleted. An error is thrown otherwise.
}
\description{
This function deletes the specified tags from a repository in GitHub.
}
\details{
\url{https://developer.github.com/v3/git/refs/#delete-a-reference}
}
