% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{delete_team}
\alias{delete_team}
\title{Delete a team in GitHub}
\usage{
delete_team(team, organization, ...)
}
\arguments{
\item{team}{(integer or string) The ID or name of the team.}

\item{organization}{(string) The login of the organization.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{delete_team()} returns a TRUE if successfully deleted.
}
\description{
This function deletes a team from an organization in GitHub, as long as you have
appropriate permissions. Care should be taken as it will not be recoverable.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/teams/#delete-team}
}
}
\examples{
\dontrun{
  # Delete a team
  delete_team("HeadCoos", "HairyCoos")
}

}
