% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{.compare_commits}
\alias{.compare_commits}
\title{View commits made between two commits}
\usage{
.compare_commits(base, head, repo, ...)
}
\arguments{
\item{base}{(string) Either a SHA, branch or tag used to identify the base commit.}

\item{head}{(string) Either a SHA, branch or tag used to identify the head commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{.compare_commits()} returns a tibble of commit properties.

\strong{Commit Properties:}
\itemize{
\item \strong{sha}: The commit SHA.
\item \strong{message}: The commit message.
\item \strong{author_login}: The author's account login.
\item \strong{author_name}: The author's name.
\item \strong{author_email}: The author's email address.
\item \strong{committer_login}: The committer's account login.
\item \strong{committer_name}: The committer's name.
\item \strong{committer_email}: The committer's email address.
\item \strong{tree_sha}: The SHA of the commit's file tree.
\item \strong{parents}: The SHAs of the parent commits.
\item \strong{date}: The time and date the commit was made.
\item \strong{html_url}: The address of the commit's web page.
}
}
\description{
\code{.compare_commits()} summarises the commits made between two commits in a table with the
properties as columns and a row for each commit. The \code{base} commit must be in the history
of the \code{head} commit.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
}
