% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{.view_branches}
\alias{.view_branches}
\alias{view_branch}
\title{View branches within a repository}
\usage{
.view_branches(repo, n_max = 1000, ...)

view_branch(branch, repo, ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}

\item{branch}{(string) The name of the branch.}
}
\value{
\code{.view_branches()} returns a tibble of branch properties. \code{view_branch()}
returns a list of properties for a single branch.

\strong{Branch Properties:}
\itemize{
\item \strong{name}: The name of the branch.
\item \strong{ref}: The full Git reference of the branch.
\item \strong{sha}: The commit SHA the branch is pointing at.
}
}
\description{
\code{.view_branches()} summarises branches in a table with the properties as columns and a row
for each branch in the repository. \code{view_branch()} returns a list of all properties for a
single branch.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#list-matching-references}
\item \url{https://developer.github.com/v3/git/refs/#get-a-single-reference}
}
}
\examples{
\dontrun{

  # View all branches in a repository
  .view_branches("ChadGoymer/githapi")

  # View a single branch
  view_label("new-branch", "ChadGoymer/githapi")

}

}
