% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{.view_releases}
\alias{.view_releases}
\alias{view_release}
\alias{browse_release}
\title{View releases within a repository}
\usage{
.view_releases(repo, n_max = 1000, ...)

view_release(release, repo, ...)

browse_release(release, repo, ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{release}{(string) The id or tag of the release.}
}
\value{
\code{.view_releases()} returns a tibble of release properties. \code{view_release()}
returns a list of properties for a single release. \code{browse_release} opens the
default browser on the release page and returns the URL.

\strong{Release Properties:}
\itemize{
\item \strong{id}: The id of the release.
\item \strong{tag}: The tags associated with the release.
\item \strong{name}: The name of the release.
\item \strong{body}: The description of the release.
\item \strong{commit}: The commit associated with the release.
\item \strong{draft}: Whether the release is draft.
\item \strong{prerelease}:  Whether it is a pre-release.
\item \strong{author_login}: The author's account login.
\item \strong{assets}: The name of the assets associated with the release.
\item \strong{html_url}: The address of the release's web page.
\item \strong{created_at}: The time and date the release was created.
\item \strong{published_at}: The time and date the release was published.
}
}
\description{
\code{.view_releases()} summarises releases in a table with the properties as columns and a row
for each release in the repository. \code{view_release()} returns a list of all properties for
a single release. \code{browse_release()} opens the web page for the release in the default
browser.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/releases/#list-releases-for-a-repository}
\item \url{https://developer.github.com/v3/repos/releases/#get-a-single-release}
\item \url{https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name}
}
}
\examples{
\dontrun{

  # View all releases in a repository
  .view_releases("ChadGoymer/githapi")

  # View a single release
  view_release("1.0.1", "ChadGoymer/githapi")

  # Browse the release web page
  browse_release("1.0.1", "ChadGoymer/githapi")

}

}
