% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{download_commit}
\alias{download_commit}
\title{Download a commit from GitHub}
\usage{
download_commit(ref, repo, path = getwd(), ...)
}
\arguments{
\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{path}{(string, optional) The path to the directory to upload. It must be readable.
Default: working directory.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{download_commit()} returns the path where the commit is downloaded to.
}
\description{
This function downloads all the files in a commit, plus any folders, into the path
specified.
}
\examples{
\dontrun{

  # Download the head of the master branch to the home directory
  download_commit(
    ref  = "master",
    repo = "ChadGoymer/githapi",
    path = "~")

}

}
