% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_file}
\alias{download_file}
\title{Download a file from GitHub}
\usage{
download_file(from_path, to_path, ref, repo, ...)
}
\arguments{
\item{from_path}{(string) The path to the file to download, within the
repository.}

\item{to_path}{(string) The path to download the file to.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{download_file()} returns the path where the file is downloaded to.
}
\description{
This function downloads a file in the specified commit to the path provided.
}
\details{
Note: The GitHub API imposes a file size limit of 100MB for this request.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/rest/reference/repos#get-repository-content}
}
}
\examples{
\dontrun{

  # Download the README file from the main branch
  download_file(
    from_path = "README.md",
    to_path   = "~/README.md",
    ref       = "main",
    repo      = "ChadGoymer/githapi"
  )

}

}
