% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{download_gist}
\alias{download_gist}
\title{Download a gist from GitHub}
\usage{
download_gist(gist, path, files, ...)
}
\arguments{
\item{gist}{(string) The ID of the gist.}

\item{path}{(string) The path to download the gist to.}

\item{files}{(character, optional) The files to download. If not specified
all the files in the gist will be downloaded.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{download_gist()} returns the path where the file is downloaded to.
}
\description{
This function downloads the files in the specified gist to the path provided.
}
\examples{
\dontrun{

  # Download all the files in a gist
  download_gist("806dca6b09a39e7b6326a0c8137583e6", "./gist")

  # Download a single file in a gist
  download_gist(
    gist  = "806dca6b09a39e7b6326a0c8137583e6",
    path  = "./gist",
    files = "helloworld.R"
  )

}

}
