% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-contents-deprecated.R
\name{files_exist}
\alias{files_exist}
\title{Determine whether files exist in the specified commit.}
\usage{
files_exist(
  paths,
  repo,
  ref,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{paths}{(string, optional) The paths to the files or directories in the repository.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. If not specified the head of the default branch is used.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A logical vector containing \code{TRUE} or \code{FALSE} for each file specified.
}
\description{
This function returns \code{TRUE} if the file exists and \code{FALSE} otherwise.
}
\details{
\url{https://developer.github.com/v3/repos/contents/#get-contents}
}
