% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_blob}
\alias{gh_blob}
\title{Get a blob}
\usage{
gh_blob(sha, repo, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{sha}{(string) SHA-1 of the blob}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A string containing the contents of the specified file.
}
\description{
url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
