% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_branch}
\alias{gh_branch}
\title{Get branch}
\usage{
gh_branch(branch, repo, token = gh_token(), api_url = gh_api_url(), ...)
}
\arguments{
\item{branch}{(string) The branch name.}

\item{repo}{(string) The repository specified as "owner/repo".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
string stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{...}{Parameters passed to \code{gh::gh())}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API_URL} or \code{https://api.github.com}.}
}
\value{
A list describing the branch (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/branches/#get-branch}
}
