% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_branches}
\alias{gh_branches}
\title{Get information about all the head commits in each branch.}
\usage{
gh_branches(repo, extended = FALSE, token = gh_token(),
  api_url = gh_api_url(), ...)
}
\arguments{
\item{repo}{(string) The repository specified as "owner/repo".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
string stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{...}{Parameters passed to \code{gh::gh())}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API_URL} or \code{https://api.github.com}.}
}
\value{
A tibble with columns "branch", "sha" and "url". When \code{extended = TRUE} the columns
  "date", "author_name", "author_email", "committer_name", "committer_email" and "message" are
  added.
}
\description{
This function returns useful information about the head commits of each branch in the
repository. If \code{extended = FALSE} a tibble with the columns "branch", "sha" and "url"
is returned with a row for each branch. If \code{extended = TRUE} the the tibble contains
the columns "branch", "sha", date", "author_name", "author_email", "committer_name",
"committer_email", "message" and "url".
}
\details{
Note: Extended results requires an separate API call for every required branch. This may
impact performance of the function considerably depending on the number of branches.
}
