% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{gh_column}
\alias{gh_column}
\title{Get a project column}
\usage{
gh_column(column, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{column}{(integer) The ID of the column in GitHub.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A list describing the column (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/projects/columns/#get-a-project-column}
}
