% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects-deprecated.R
\name{gh_columns}
\alias{gh_columns}
\title{List project columns}
\usage{
gh_columns(
  project,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{project}{(integer) The ID of the project in GitHub.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the columns (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/projects/columns/#list-project-columns}
}
