% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-deprecated.R
\name{gh_commit_comments}
\alias{gh_commit_comments}
\title{List comments for a single commit or entire repository}
\usage{
gh_commit_comments(
  repo,
  ref,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch is
specified the head commit is used.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing all the commit comments (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/comments/#list-comments-for-a-single-commit}
\url{https://developer.github.com/v3/repos/comments/#list-commit-comments-for-a-repository}
}
