% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_compare_commits}
\alias{gh_compare_commits}
\title{Compare two commits}
\usage{
gh_compare_commits(base, head, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{base}{(string) The base branch name.}

\item{head}{(string) The branch to compare to the base.}

\item{repo}{(string) The repository specified as "owner/repo".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A list describing the commits and the differences (see GitHub's API documentation for
  details).
}
\description{
\url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
