% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-deprecated.R
\name{gh_compare_files}
\alias{gh_compare_files}
\title{Compare the files of two commits}
\usage{
gh_compare_files(
  base,
  head,
  repo,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{base}{(string) The base branch name.}

\item{head}{(string) The branch to compare to the base.}

\item{repo}{(string) The repository specified as "owner/repo".}

\item{token}{(string, optional) The personal access token for GitHub authorisation.
Default: value stored in the environment variable \code{GITHUB_TOKEN} or
\code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A tibble describing the files and the differences (see GitHub's API documentation
for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=compare_files]{compare_files()}} instead.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
