% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-deprecated.R
\name{gh_contributers}
\alias{gh_contributers}
\title{List contributors}
\usage{
gh_contributers(
  repo,
  anon = NULL,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{anon}{(boolean) Set to TRUE to include anonymous contributors in results.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the contributers (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/#list-contributors}
}
