% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_download}
\alias{gh_download}
\title{Download the entire tree from a commit}
\usage{
gh_download(ref, repo, path, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{ref}{(string) A git reference: either a SHA-1, tag or branch. If a branch is specified
the head commit is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{path}{(string) The path to save to.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_download_binary]{gh_download_binary()}}.}
}
\value{
Full tree of files in specified location.
}
\description{
Download the entire tree from a commit
}
