% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_download_binary}
\alias{gh_download_binary}
\title{Download a binary file from GitHub}
\usage{
gh_download_binary(url, path, token = getOption("github.token"))
}
\arguments{
\item{url}{(string) URL to the GitHub API.}

\item{path}{(string) The location to download the file to.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}
}
\value{
The path to the download location
}
\description{
Download a binary file from GitHub
}
