% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gh_download_gist}
\alias{gh_download_gist}
\title{Download files in a gist}
\usage{
gh_download_gist(gist, save_to, files, sha, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{gist}{(string) The ID of the gist in GitHub.}

\item{save_to}{(string) The file path to save the file to.}

\item{files}{(character, optional) The name of the files to download. Default: all files in
the gist}

\item{sha}{(string, optional) The SHA-1 of the version required. Default: latest version.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
The file path of the saved file (invisibly).
}
\description{
url{https://developer.github.com/v3/gists/#get-a-single-gist}
url{https://developer.github.com/v3/gists/#get-a-specific-revision-of-a-gist}
}
