% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_event}
\alias{gh_event}
\title{Get a single event}
\usage{
gh_event(event, repo, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{event}{(integer) The ID assigned to the event in GitHub.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the event (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/events/#get-a-single-event}
}
