% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_events}
\alias{gh_events}
\title{List events for an issue or all issues in a repository}
\usage{
gh_events(repo, issue, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{issue}{(integer, optional) The number assigned to the issue.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A tibble describing the issue events (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/events/#list-events-for-a-repository}
\url{https://developer.github.com/v3/issues/events/#list-events-for-an-issue}
}
