% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{gh_find}
\alias{gh_find}
\title{Find an entity by matching a property value}
\usage{
gh_find(url, property, value, max_pages = 100, headers = NULL,
  accept = "application/vnd.github.v3+json", token = gh_token(),
  proxy = getOption("github.proxy"))
}
\arguments{
\item{url}{(string) The address of the API endpoint.}

\item{property}{(string) The property to search.}

\item{value}{(string) The property value to search for.}

\item{max_pages}{(integer, optional) The maximum number of pages to search through.
Default: 100.}

\item{headers}{(character, optional) Headers to add to the request. Default: \code{NULL}.}

\item{accept}{(string, optional) The mime format to accept when making the call. Default:
"application/vnd.github.v3+json".}

\item{token}{(string or Token, optional) An authorisation token to include with the
request. If \code{NULL} the OAuth process is triggered. Default: \code{NULL}.}

\item{proxy}{(character, optional) The proxy server to use to connect to the github API.
If \code{NULL} then no proxy is used. Can be set in the option \code{github.proxy} or the
environment variable \code{GITHUB_PROXY}. Default: \code{NULL}.}
}
\value{
A \code{github} list object consisting of the response, parsed into a list, with the
attributes:
\itemize{
\item \strong{url}: The URLs the request was sent to
\item \strong{request}: The type of HTTP request made
\item \strong{status}: The HTTP status code returned
\item \strong{header}: The HTTP header returned
}
}
\description{
This function pages through a collection of entities searching for a specified property
value. It returns the first match found. For example, you can search for an issue by
specifying the title.
}
\examples{
\dontrun{

  # Find an issue by title
  gh_find(
    url      = "https://api.github.com/repos/ChadGoymer/githapi/issues",
    property = "title",
    value    = "Test issue")
}

}
