% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_get}
\alias{gh_get}
\title{Get the contents of a github http request as a string.}
\usage{
gh_get(address, binary = FALSE, github_url = getOption("github.url"),
  token = gh_token())
}
\description{
Get the contents of a github http request as a string.
}
