% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gh_gist}
\alias{gh_gist}
\title{Get a single gist}
\usage{
gh_gist(gist, sha, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A list describing the gist (see GitHub's API documentation for details).
}
