% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gh_gist_comment}
\alias{gh_gist_comment}
\title{Get a single comment}
\usage{
gh_gist_comment(comment, gist, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{comment}{(integer) The ID of the comment in GitHub.}

\item{gist}{(string) The ID of the gist in GitHub.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the gist comment (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/gists/comments/#get-a-single-comment}
}
