% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gh_gist_forks}
\alias{gh_gist_forks}
\title{List gist forks}
\usage{
gh_gist_forks(gist, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{gist}{(string) The ID of the gist in GitHub.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A tibble describing the forks (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/gists/#list-gist-forks}
}
