% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_git_blob}
\alias{gh_git_blob}
\title{Get a blob}
\usage{
gh_git_blob(sha, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{sha}{(string) SHA-1 of the blob}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A string containing the contents of the specified file (see GitHub's API
documentation for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=view_blobs]{view_blobs()}} instead.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
