% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_git_commit}
\alias{gh_git_commit}
\title{Get a commit}
\usage{
gh_git_commit(sha, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{sha}{(string) The SHA-1 of the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the commit (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/git/commits/#get-a-commit}
}
