% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_git_tag}
\alias{gh_git_tag}
\title{Get an annotated tag}
\usage{
gh_git_tag(sha, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{sha}{(string) The SHA-1 of the tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the tag (see GitHub's API documentation for details).
}
\description{
NOTE: This function has been deprecated, it will be removed in a future version.
}
\details{
\url{https://developer.github.com/v3/git/tags/#get-a-tag}
}
