% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_git_tree}
\alias{gh_git_tree}
\title{Get a tree}
\usage{
gh_git_tree(ref, repo, recursive = TRUE, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{ref}{(string) A git reference: either a SHA-1, tag or branch. If a branch is specified
the head commit is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{recursive}{(logical) Whether to list files recursively. Default: TRUE.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A tibble describing the files in a commit (see GitHub's API documentation for
details).
}
\description{
\url{https://developer.github.com/v3/git/trees/#get-a-tree}
}
