% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_issue}
\alias{gh_issue}
\title{Get a single issue}
\usage{
gh_issue(issue_number, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{issue_number}{(integer) The number assigned to the issue.}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A list describing the issue (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/#get-a-single-issue}
}
