% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_issue_comment}
\alias{gh_issue_comment}
\title{Get a single issue comment}
\usage{
gh_issue_comment(comment_id, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{comment_id}{(integer) The number assigned to the comment.}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A list describing the issue comment (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/comments/#get-a-single-comment}
}
