% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_issues}
\alias{gh_issues}
\title{List issues for a repository}
\usage{
gh_issues(repo, milestone = NULL, state = NULL, assignee = NULL,
  creator = NULL, mentioned = NULL, labels = NULL, sort = NULL,
  direction = NULL, since = NULL, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{milestone}{(string)    If an integer is passed, it should refer to a milestone by its number
field. If the string * is passed, issues with any milestone are accepted. If the string none
is passed, issues without milestones are returned.}

\item{state}{(string) Indicates the state of the issues to return. Can be either open, closed,
or all. Default: open}

\item{assignee}{(string) Can be the name of a user. Pass in none for issues with no assigned
user, and * for issues assigned to any user.}

\item{creator}{(string) The user that created the issue.}

\item{mentioned}{(string) A user that's mentioned in the issue.}

\item{labels}{(string) A list of comma separated label names. Example: bug,ui,@high}

\item{sort}{(string) What to sort results by. Can be either created, updated, comments.
Default: created}

\item{direction}{(string) The direction of the sort. Can be either asc or desc. Default: desc}

\item{since}{(string) Only issues updated at or after this time are returned. This is a
timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{n_max}{(integer) Maximum number to return.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing all the issues a repository has (see GitHub's API documentation for
  more detail)
}
\description{
\url{https://developer.github.com/v3/issues/#list-issues-for-a-repository}
}
