% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_json}
\alias{gh_json}
\title{Get the JSON contents of a GitHub request and parse into a list.}
\usage{
gh_json(address, token = gh_token())
}
\description{
Get the JSON contents of a GitHub request and parse into a list.
}
