% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_labels}
\alias{gh_labels}
\title{List all labels for this repository}
\usage{
gh_labels(repo, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A list describing the label (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository}
}
