% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{gh_members}
\alias{gh_members}
\title{Get Organization or team members list}
\usage{
gh_members(org, team, filter = NULL, role = NULL, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{org}{(string) The name of the organization.}

\item{team}{(integer) The GitHub ID of the team.}

\item{filter}{(string) Filter members returned in the list. Can be one of:
\itemize{
  \item 2fa_disabled: Members without two-factor authentication enabled. Available for
    organization owners.
  \item all: All members the authenticated user can see.
  \item Default: all.
}}

\item{role}{(string) Filter members returned by their role. Can be one of:
\itemize{
  \item all: All members of the organization, regardless of role.
  \item admin: Organization owners.
  \item member: Non-owner organization members.
  \item Default: all.
}}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing the members (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/orgs/members/#members-list}
url{https://developer.github.com/v3/orgs/teams/#list-team-members}
}
