% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{gh_memberships}
\alias{gh_memberships}
\title{List your organization memberships}
\usage{
gh_memberships(org, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{org}{(string, optional) The name of the organization.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing your memberships (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/orgs/members/#list-your-organization-memberships}
url{https://developer.github.com/v3/orgs/members/#get-your-organization-membership}
}
