% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_milestones}
\alias{gh_milestones}
\title{List milestones for a repository}
\usage{
gh_milestones(repo, state = NULL, sort = NULL, direction = NULL,
  n_max = 1000L, token = gh_token(), api = getOption("github.api"),
  ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{state}{(string) The state of the milestone. Either open, closed, or all. Default: open}

\item{sort}{(string) What to sort results by. Either due_on or completeness. Default: due_on}

\item{direction}{(string) The direction of the sort. Either asc or desc. Default: asc}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the milestones (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository}
}
