% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{gh_organization}
\alias{gh_organization}
\title{Get an organization}
\usage{
gh_organization(org, token = gh_token(), api = getOption("github.api"),
  ...)
}
\arguments{
\item{org}{(string) The name of the organization.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the organization (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/orgs/#get-an-organization}
}
