% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_page}
\alias{gh_page}
\title{Get and parse the contents of a github http request, including multiple pages.}
\usage{
gh_page(address, ..., n_max = 1000L, token = gh_token())
}
\description{
Get and parse the contents of a github http request, including multiple pages.
}
