% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_permissions}
\alias{gh_permissions}
\title{Review a user's permission level}
\usage{
gh_permissions(user, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{user}{(string) The GitHub username of the user.}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A list describing the user's permissions (see GitHub's API documentation for
  details).
}
\description{
url{https://developer.github.com/v3/repos/collaborators/#review-a-users-permission-level}
}
