% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects-deprecated.R
\name{gh_project}
\alias{gh_project}
\title{Get a project}
\usage{
gh_project(project, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{project}{(integer) The ID of the project in GitHub.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the project (see GitHub's API documentation for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=view_project]{view_project()}} instead.
}
\details{
\url{https://developer.github.com/v3/projects/#get-a-project}
}
