% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects-deprecated.R
\name{gh_projects}
\alias{gh_projects}
\title{List organization or repository projects}
\usage{
gh_projects(
  repo,
  org,
  state = NULL,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{org}{(string) The name of the organization.}

\item{state}{(string, optional) Indicates the state of the projects to return. Can be either open,
closed, or all. Default: open}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the projects (see GitHub's API documentation for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=view_projects]{view_projects()}} instead.
}
\details{
\url{https://developer.github.com/v3/projects/#list-repository-projects}
\url{https://developer.github.com/v3/projects/#list-organization-projects}
}
