% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_pull_comments}
\alias{gh_pull_comments}
\title{List review comments on a pull request or all pull requests in a repository.}
\usage{
gh_pull_comments(repo, pull_request, review, sort = NULL, direction = NULL,
  since = NULL, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{pull_request}{(integer, optional) The number assigned to the pull request. If
not specified, the comments on all pull requests are listed.}

\item{review}{(integer, optional) The number assigned to the review. If specified,
along with a \code{pull_request}, only the comments on this review are returned.}

\item{sort}{(string) Can be either created or updated. Default: created}

\item{direction}{(string) Can be either asc or desc. Ignored without sort parameter.}

\item{since}{(string) Only comments updated at or after this time are returned. This is a
timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ.}

\item{n_max}{(integer) Maximum number to return.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing the pull request comments (see GitHub's API documentation for
  details).
}
\description{
url{https://developer.github.com/v3/pulls/reviews/#get-comments-for-a-single-review}
url{https://developer.github.com/v3/pulls/comments/#list-comments-on-a-pull-request}
url{https://developer.github.com/v3/pulls/comments/#list-comments-in-a-repository}
}
