% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_pull_review}
\alias{gh_pull_review}
\title{Get a single review}
\usage{
gh_pull_review(
  review,
  pull_request,
  repo,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{review}{(integer) The number assigned to the review.}

\item{pull_request}{(integer) The number assigned to the pull request.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the pull request review (see GitHub's API documentation for
details).
}
\description{
\url{https://developer.github.com/v3/pulls/reviews/#get-a-single-review}
}
