% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_pull_review_requests}
\alias{gh_pull_review_requests}
\title{List pull request review requests}
\usage{
gh_pull_review_requests(
  pull_request,
  repo,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{pull_request}{(integer) The number assigned to the pull request.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the pull request review requests (see GitHub's API
documentation for details).
}
\description{
\url{https://developer.github.com/v3/pulls/review_requests/#list-review-requests}
}
