% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_release}
\alias{gh_release}
\title{Get a single release}
\usage{
gh_release(tag, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{tag}{(string, optional) The tag associated with a release. If not supplied the latest
release is returned.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the release (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/releases/#get-a-single-release}
\url{https://developer.github.com/v3/repos/releases/#get-the-latest-release}
\url{https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name}
}
