% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_repos}
\alias{gh_repos}
\title{Get information about a user or organisation's repositories.}
\usage{
gh_repos(owner, type = "owner", sort = "full_name",
  direction = ifelse(sort == "full_name", "asc", "desc"), limit = Inf,
  token = gh_token(), api = gh_api(), ...)
}
\arguments{
\item{owner}{(string) The user or organisation owning the repository.}

\item{type}{(string, optional) Can be one of \code{"all"}, \code{"owner"}, \code{"member"}.
Default: \code{"owner"}.}

\item{sort}{(string, optional) Can be one of \code{"created"}, \code{"updated"},
\code{"pushed"}, \code{"full_name"}. Default: \code{"full_name"}.}

\item{direction}{(string, optional) Can be one of \code{"asc"} or \code{"desc"}. Default:
when using \code{"full_name"}: \code{"asc"}, otherwise \code{"desc"}.}

\item{limit}{(integer, optional) The maximum number to return. Default: \code{Inf}, which
returns all.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh}}.}
}
\value{
A tibble describing all the repositories a user or organisation has (see GitHub's
  API documentation for more detail).
}
\description{
\url{https://developer.github.com/v3/repos/#list-user-repositories}
\url{https://developer.github.com/v3/repos/#list-organization-repositories}
}
