% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_request}
\alias{gh_request}
\title{Send a http request to the specified GitHub url}
\usage{
gh_request(type, url, payload = NULL, accept = "json", parse = TRUE,
  token = gh_token())
}
\arguments{
\item{type}{(string) The type of the request. Either "GET", "POST", "DELETE" or "PATCH".}

\item{url}{(string) URL to the GitHub API.}

\item{payload}{(list) The information to send, which is converted to JSON.}

\item{accept}{(string) The format of the returned result. Either "json", "raw" or other
GitHub accepted format. Default: "json".}

\item{parse}{(boolean) Whether to parse the response.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}
}
\value{
A list of the parsed response
}
\description{
Send a http request to the specified GitHub url
}
