% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_save}
\alias{gh_save}
\title{Download file and save it}
\usage{
gh_save(path, repo, save_to = basename(path), ref = "master",
  token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{path}{(string) The path to the file in the repository}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{save_to}{(string, optional) The file path to save the file to. Default: file name
specified in \code{path} in the current working directory.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. Default: "master".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A string containing the contents of the specified file (see GitHub's API
  documentation for details).
}
\description{
url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
