% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists-deprecated.R
\name{gh_source_gist}
\alias{gh_source_gist}
\title{Source an R file from a gist}
\usage{
gh_source_gist(
  file,
  gist,
  sha,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{file}{(string) The name of the file to source.}

\item{gist}{(string) The ID of the gist in GitHub.}

\item{sha}{(string, optional) The SHA-1 of the version required. Default: latest version.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[base:source]{base::source()}}.}
}
\value{
Nothing. The file is sourced into global environment.
}
\description{
\url{https://developer.github.com/v3/gists/#get-a-single-gist}
\url{https://developer.github.com/v3/gists/#get-a-specific-revision-of-a-gist}
}
