% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-deprecated.R
\name{gh_tags}
\alias{gh_tags}
\title{Get information about all the tags in a repository}
\usage{
gh_tags(
  repo,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing all the tags (see GitHub's API documentation for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=view_tags]{view_tags()}} instead.
}
\details{
\url{https://developer.github.com/v3/repos/#list-tags}
}
