% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations-deprecated.R
\name{gh_team}
\alias{gh_team}
\title{Get team}
\usage{
gh_team(team, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{team}{(integer) The GitHub ID of the team.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the team (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/orgs/teams/#get-team}
}
