% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{gh_teams}
\alias{gh_teams}
\title{List teams}
\usage{
gh_teams(org, repo, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{org}{(string, optional) The name of the organization. If org and repo are not
specified, the teams of the authenticated user are returned.}

\item{repo}{(string, optional) The repository specified in the format: \code{"owner/repo"}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing the teams (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/orgs/teams/#list-teams}
url{https://developer.github.com/v3/orgs/teams/#list-user-teams}
}
