% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_tibble}
\alias{gh_tibble}
\title{Get and parse the contents of a github http request, including multiple pages and return a
tibble.}
\usage{
gh_tibble(address, ..., n_max = 1000L, token = gh_token())
}
\description{
Get and parse the contents of a github http request, including multiple pages and return a
tibble.
}
