% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi-deprecated.R
\name{gh_token}
\alias{gh_token}
\title{Retrieve the GitHub personal access token}
\usage{
gh_token(github_token = getOption("github.token"),
  github_oauth_url = getOption("github.oauth_url"),
  githapi_key = getOption("githapi.key"),
  githapi_secret = getOption("githapi.secret"),
  githapi_cache = getOption("githapi.cache"))
}
\value{
The token as a string
}
\description{
The token is retreived by looking in the following locations, in order:
}
\details{
\enumerate{
\item \code{github.token} option
\item \code{GITHUB_TOKEN} environment variable
\item \code{GITHUB_PAT} environment variable
}
}
