% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_url}
\alias{gh_url}
\title{Build the URL for the github API}
\usage{
gh_url(..., api = getOption("github.api"))
}
\arguments{
\item{...}{(strings, optional) unnamed strings are built up into a URL path and named
parameters are added as queries.}

\item{api}{(string, optional) the base address of GitHub's API. Default:
\code{getOption("github.api")}}
}
\value{
Valid URL (string)
}
\description{
Build the URL for the github API
}
