% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{gh_user}
\alias{gh_user}
\title{Get a single user}
\usage{
gh_user(user, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{user}{(string) The GitHub username of the user.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh}}.}
}
\value{
A list describing the user (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/users/#get-a-single-user}
}
