% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users-deprecated.R
\name{gh_user_email}
\alias{gh_user_email}
\title{List email addresses for the authenticated user}
\usage{
gh_user_email(token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
string stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A tibble describing the user's email addresses (see GitHub's API documentation for
more detail).
}
\description{
\url{https://developer.github.com/v3/users/emails/#list-email-addresses-for-a-user}
\url{https://developer.github.com/v3/users/emails/#list-public-email-addresses-for-a-user}
}
