% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{gh_users}
\alias{gh_users}
\title{Get all users}
\usage{
gh_users(token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
string stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API_URL} or \code{https://api.github.com}.}

\item{limit}{(integer, optional) The maximum number to return. Default: 1000.}
}
\value{
A tibble describing the users (see GitHub's API documentation for more detail).
}
\description{
\url{https://developer.github.com/v3/users/#get-all-users}
}
