% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi-deprecated.R
\name{githapi-deprecated}
\alias{githapi-deprecated}
\title{Deprecated functions in package \link{githapi}.}
\description{
The functions listed below are deprecated and will be removed in the near future. When
possible, alternative functions with similar functionality are also mentioned.
}
\section{Git Data Functions}{

\itemize{
\item \code{\link[=gh_git_blob]{gh_git_blob()}}: use \code{\link[=view_blobs]{view_blobs()}} instead.
\item \code{\link[=gh_git_commit]{gh_git_commit()}}: use \code{\link[=view_commits]{view_commits()}} instead.
\item \code{\link[=gh_git_reference]{gh_git_reference()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=gh_git_references]{gh_git_references()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=is_tag]{is_tag()}}: use \code{\link[=tags_exist]{tags_exist()}} instead.
\item \code{\link[=gh_git_tag]{gh_git_tag()}}: is being removed
\item \code{\link[=gh_git_tree]{gh_git_tree()}}: replace with \code{\link[=view_trees]{view_trees()}}
\item \code{\link[=gh_save]{gh_save()}}: replace with \code{\link[=download_files]{download_files()}}
\item \code{\link[=gh_source]{gh_source()}}: replace with \code{\link[=source_files]{source_files()}}
}
}

\section{Repositories Functions}{

\itemize{
\item \code{\link[=gh_tags]{gh_tags()}}: replace with \code{\link[=view_tags]{view_tags()}}
\item \code{\link[=gh_branch]{gh_branch()}}: replace with \code{\link[=view_branches]{view_branches()}}
\item \code{\link[=gh_branches]{gh_branches()}}: replace with \code{\link[=view_branches]{view_branches()}}
}
}

\keyword{internal}
