% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\docType{package}
\name{githapi}
\alias{githapi}
\alias{githapi-package}
\title{User-friendly access to the GitHub API for R, consistent with the tidyverse.}
\description{
Provides a suite of functions which simplify working with GitHub's API.
}
