% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{github_source}
\alias{github_source}
\title{Source a R script from a commit}
\usage{
github_source(path, ref, repo, ...)
}
\arguments{
\item{path}{(string) The path to the file, within the repository.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=source]{source()}}.}
}
\value{
The result of the sourced script.
}
\description{
This function sources an R script from a commit in a repository using \code{\link[=source]{source()}}.
}
\examples{
\dontrun{

  github_source(
    path = "inst/test-data/test-script.R",
    ref  = "master",
    repo = "ChadGoymer/githapi")

}

}
