% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{is_branch}
\alias{is_branch}
\title{Checks whether the input is a valid branch.}
\usage{
is_branch(branch, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{branch}{(string) The branch name.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
TRUE or FALSE.
}
\description{
\url{https://developer.github.com/v3/repos/branches/#get-branch}
}
