% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{is_collaborator}
\alias{is_collaborator}
\title{Check if a user is a collaborator}
\usage{
is_collaborator(user, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{user}{(string) The GitHub username of the user.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
TRUE if the user is a collaborator, FALSE otherwise (see GitHub's API
documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/collaborators/#check-if-a-user-is-a-collaborator}
}
