% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{is_gist_starred}
\alias{is_gist_starred}
\title{Check if a gist is starred}
\usage{
is_gist_starred(gist, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{gist}{(string) The ID of the gist in GitHub.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
TRUE if the gist has been starred, FALSE otherwise (see GitHub's API documentation
for details).
}
\description{
\url{https://developer.github.com/v3/gists/#check-if-a-gist-is-starred}
}
