% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{is_member}
\alias{is_member}
\title{Check membership}
\usage{
is_member(user, org, token = gh_token(), api = getOption("github.api"),
  ...)
}
\arguments{
\item{user}{(string) The GitHub username of the user.}

\item{org}{(string) The name of the organization.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
TRUE if the user is a member, FALSE otherwise (see GitHub's API documentation for
details).
}
\description{
\url{https://developer.github.com/v3/orgs/members/#check-membership}
}
