% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{is_repo}
\alias{is_repo}
\title{Checks whether the supplied object is a valid repo name}
\usage{
is_repo(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a valid repo, FALSE otherwise
}
\description{
Checks whether the supplied object is a valid repo name
}
