% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{is_repository}
\alias{is_repository}
\title{Check whether the input is a valid repository}
\usage{
is_repository(repo, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_get}}.}
}
\value{
A boolean, with attributes describing the errors, if there are any.
}
\description{
\url{https://developer.github.com/v3/repos/#get}
}
