% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{is_tag}
\alias{is_tag}
\title{Check whether the input is a valid tag.}
\usage{
is_tag(tag, repo, token = gh_token(), api = getOption("github.api"), ...)
}
\arguments{
\item{tag}{(string) tag name.}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A boolean, with attributes describing the errors, if there are any.
}
\description{
url{https://developer.github.com/v3/git/refs/#get-a-reference}
}
