% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{is_url}
\alias{is_url}
\title{Checks whether the supplied object is a valid URL}
\usage{
is_url(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a valid URL, FALSE otherwise
}
\description{
Checks whether the supplied object is a valid URL
}
