% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{is_valid_sha}
\alias{is_valid_sha}
\title{Check whether the input is a valid SHA-1}
\usage{
is_valid_sha(sha, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{sha}{(string) A commit SHA-1.}

\item{repo}{(string, optional) The repository specified in the format: \code{owner/repo}.
If not supplied existence of the commit is not checked, only the format of the string.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
TRUE or FALSE.
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=shas_exist]{shas_exist()}} instead.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#get-a-single-commit}
}
