% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-blobs.R
\name{read_files}
\alias{read_files}
\title{Read the contents of text files}
\usage{
read_files(
  paths,
  ref,
  repo,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{paths}{(string) The paths to the files.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. If not specified the head of the default branch is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A named character vector with each element containing a file's contents.
}
\description{
This function reads the contents of the specified files from a repository in GitHub.
Note: This API supports files up to 100 megabytes in size.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
