% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-releases.R
\name{releases_exist}
\alias{releases_exist}
\title{Determine whether releases exist in the specified repository.}
\usage{
releases_exist(tags, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{tags}{(character) The names of the release tags.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A logical vector containing \code{TRUE} or \code{FALSE} for each tag specified.
}
\description{
This function returns \code{TRUE} if the release exists and \code{FALSE} otherwise.
}
\details{
\url{https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name}
}
