% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{source_gist}
\alias{source_gist}
\title{Source a gist from GitHub}
\usage{
source_gist(gist, files, ...)
}
\arguments{
\item{gist}{(string) The ID of the gist.}

\item{files}{(character, optional) The files to source. If not specified all the files
in the gist will be sourced.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{source_gist()} returns the result of sourcing the files.
}
\description{
This function downloads the files in the specified gist and then sources them using the
\code{\link[=source]{source()}} function.
}
\examples{
\dontrun{

  # Source all the files in a gist
  source_gist("806dca6b09a39e7b6326a0c8137583e6", "./gist")

  # Source a single file in a gist
  source_gist(
    gist  = "806dca6b09a39e7b6326a0c8137583e6",
    path  = "./gist",
    files = "helloworld.R")

}

}
