% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{update_branch}
\alias{update_branch}
\title{Update a branch in a repository}
\usage{
update_branch(branch, ref, repo, force = FALSE, ...)
}
\arguments{
\item{branch}{(string) The name of the branch.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the new commit the branch
is pointing at.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{force}{(boolean, optional) Whether to force the update if it is not a simple
fast-forward. Default: \code{FALSE}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_branch()} returns a list of the branch properties.

\strong{Branch Properties:}
\itemize{
\item \strong{name}: The name of the branch.
\item \strong{ref}: The full Git reference of the branch.
\item \strong{sha}: The commit SHA the branch is pointing at.
}
}
\description{
This function updates a branch in the specified repository to point at a new commit. It
must be pointed at a commit by providing a Git reference, which can be either a SHA, branch
or tag. For a branch, the head commit is used.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#update-a-reference}
}
}
\examples{
\dontrun{
  update_branch(
    name = "new-branch",
    repo = "ChadGoymer/test-githapi",
    ref  = "6b7b5a090d47fd3ef495620513a3f80da2487b1d")
}

}
