% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branches.R
\name{update_branches}
\alias{update_branches}
\title{Update branches to new commits.}
\usage{
update_branches(branches, shas, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{branches}{(character) The branch names.}

\item{shas}{(character) The SHAs of the commits to update the branch to.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the branches, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{name}: The name of the branch.
\item \strong{ref}: The full git reference for the branch.
\item \strong{url}: The URL to get the branch details from GitHub.
\item \strong{object_sha}: The SHA of the object at the HEAD of the branch.
\item \strong{object_type}: The type of the object at the HEAD of the branch.
\item \strong{object_url}: The URL to get the object details from GitHub.
}
}
\description{
This function updates the specified branches to point at new commits within a repository in
GitHub.
}
\details{
\url{https://developer.github.com/v3/git/refs/#update-a-reference}
}
