% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{update_card}
\alias{update_card}
\alias{move_card}
\title{Update a card in a GitHub project}
\usage{
update_card(card, note, archived, ...)

move_card(card, position, after, column, project, repo, user, org, ...)
}
\arguments{
\item{card}{(integer) The ID of the card.}

\item{note}{(string, optional) The new note for the card.}

\item{archived}{(boolean, optional) Whether to archive the card.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}

\item{position}{(string, optional) Either \code{"top"} or \code{"bottom"}.}

\item{after}{(integer, optional) An ID of another card to place this one
after.}

\item{column}{(integer or string, optional) Either the ID or name of the
column.}

\item{project}{(integer or string) Either the project number or name. Only
required if moving the column.}

\item{repo}{(string, optional) The repository specified in the format:
\code{owner/repo}.}

\item{user}{(string, optional) The login of the user.}

\item{org}{(string, optional) The name of the organization.}
}
\value{
\code{update_card()} returns a list of the card properties.

\strong{Card Properties:}
\itemize{
\item \strong{id}: The ID of the card.
\item \strong{content_id}: The ID of the issue or pull request.
\item \strong{note}: The content of a note.
\item \strong{archived}: Whether the card has been archived.
\item \strong{creator}: The creator of the note.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
\code{update_card()} can be used to change a card's note in a project in GitHub or
archive it. \code{move_card()} can be used to reorder the cards or move them to
other columns.
}
\details{
You can update a card associated with either a repository, user or
organization, by supplying them as an input, as long as you have appropriate
permissions.

You can move a card by either specifying the position, either \code{"top"} or
\code{"bottom"}, by specifying another card to place it after, or by specifying a
column to move it to.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/rest/reference/projects#update-a-project-card}
\item \url{https://docs.github.com/en/rest/reference/projects#move-a-project-card}
}
}
\examples{
\dontrun{

  # Update a note in a card
  update_card(card = 123456, note = "This is an updated note")

  # Archive a card
  update_card(card = 654321, archived = TRUE)

  # Move a card to the top of a column
  move_card(card = 123456, position = "top")

  # Move a card after another card
  move_card(card  = 123456, after = 654321)

  # Move card to another column
  move_card(
    card     = 123456,
    position = "top",
    column   = "Test column 2",
    project  = "Test project",
    repo     = "ChadGoymer/githapi"
  )

}

}
