% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{update_gist}
\alias{update_gist}
\title{Update a gist in GitHub}
\usage{
update_gist(gist, files, description, ...)
}
\arguments{
\item{gist}{(string) The ID of the gist}

\item{files}{(list, optional) The file contents with the list names as the filenames.
Added a \code{filename} element changes the filename (see examples).}

\item{description}{(string, optional) A description for the gist.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_gist()} returns a list of the gist's properties.

\strong{Gist Properties:}
\itemize{
\item \strong{id}: The id for the gist
\item \strong{description}: The description of the gist
\item \strong{files}: A tibble of file properties:
\itemize{
\item \strong{filename}: The name of the file
\item \strong{type}: The type of file
\item \strong{content}: The file content
\item \strong{size}: The size of the file in bytes
\item \strong{truncated}: Whether the file content has been truncated
}
\item \strong{owner}: The login if the gist's owner
\item \strong{public}: Whether the gist is public
\item \strong{html_url}: The address of the gist's web page.
\item \strong{created_at}: The time and date the gist was created.
\item \strong{updated_at}: The time and date the gist was last updated.
}
}
\description{
This function updates an existing gist in GitHub. You can specify add or update files by
providing a named list and rename a file (see examples).
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/gists/#update-a-gist}
}
}
\examples{
\dontrun{

  # Update a gist's description
  update_gist(
    gist        = "806dca6b09a39e7b6326a0c8137583e6",
    description = "An updated description")

  # Update the contents of a file
  update_gist(files = list(hello_world.R = "cat(\"Hello World!\")"))

  # Update the contents of a file and the filename
  update_gist(
    files = list(hello_world.R = c("cat(\"Hello World!\")", filename = "new_filename.R")))

}

}
