% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{update_issue}
\alias{update_issue}
\title{Update an issue in a repository}
\usage{
update_issue(
  issue,
  repo,
  title,
  body,
  assignees,
  labels,
  milestone,
  state,
  ...
)
}
\arguments{
\item{issue}{(string or character) The number or title of the issue.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{title}{(string, optional) The title of the issue.}

\item{body}{(string, optional) The contents of the issue.}

\item{assignees}{(character, optional) Logins for Users to assign to this issue. NOTE: Only
users with push access can set assignees for new issues. Setting to \code{NULL} clears all
assignees.}

\item{labels}{(character, optional) Labels to associate with this issue. NOTE: Only users
with push access can set labels for new issues. Setting to \code{NULL} clears all labels.}

\item{milestone}{(character or integer, optional) The title or number of the milestone to
associate this issue with. NOTE: Only users with push access can set the milestone for
new issues. Setting to \code{NULL} clears the current milestone.}

\item{state}{(string, optional) The state of the issue. Either \code{"open"} or \code{"closed"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_issue()} returns a list of the issue's properties.

\strong{Issue Properties:}
\itemize{
\item \strong{number}: The number assigned to the issue.
\item \strong{title}: The title of the issue.
\item \strong{body}: The body contents of the issue.
\item \strong{assignees}: The users assigned to the issue.
\item \strong{labels}: The labels attached to the issue.
\item \strong{milestone}: The milestone assigned to the issue.
\item \strong{state}: The state of the issue - either \code{"open"} or \code{"closed"}.
\item \strong{repository}: The repository the issue is in.
\item \strong{pull_request}: Whether the issue is a pull request.
\item \strong{html_url}: The URL of the issue's web page in GitHub.
\item \strong{creator}: The creator's login.
\item \strong{created_at}: When the issue was created.
\item \strong{updated_at}: When the issue was last updated.
\item \strong{closed_at}: When the issue was closed.
}
}
\description{
This function updates an issue for the specified repository in GitHub. It can be used to
change the title, body, or assignees, it can also be used to replace labels and milestones,
or to close the issue.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/#edit-an-issue}
}
}
\examples{
\dontrun{

  # Update the properties of a issue
  update_issue(
    issue   = "test issue",
    repo        = "ChadGoymer/test-githapi",
    title       = "updated test issue",
    description = "This is an updated test issue",
    due_on      = "2020-12-01")

  # Close a issue
  update_issue(
    issue = "updated test issue",
    repo      = "ChadGoymer/test-githapi",
    state     = "closed")

}

}
