% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{update_label}
\alias{update_label}
\title{Update a label in a repository}
\usage{
update_label(label, repo, name, color, description, ...)
}
\arguments{
\item{label}{(string) The name of the label.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{name}{(string, optional) The name of the label.}

\item{color}{(string, optional) Either the color name (see \code{\link[grDevices:colors]{grDevices::colors()}}) or a
hexidecimal color code (see \href{http://www.color-hex.com/}{color-hex.com}).}

\item{description}{(string, optional) A description of the label.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_label()} returns a list of the label properties.

\strong{Label Properties:}
\itemize{
\item \strong{name}: The name of the label.
\item \strong{color}: The color of the label in hexidecimal code.
\item \strong{description}: The description of the label.
}
}
\description{
This function updates a label for the specified repository in GitHub. It can be used to
change name, color or description.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/labels/#update-a-label}
}
}
\examples{
\dontrun{

  update_label(
    label       = "new-label",
    repo        = "ChadGoymer/githapi",
    name        = "updated-label",
    color       = "green",
    description = "This is an updated label")

}

}
