% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{update_membership}
\alias{update_membership}
\title{Update membership of an organization or team}
\usage{
update_membership(user, org, team, role, ...)
}
\arguments{
\item{user}{(string) The login of the user.}

\item{org}{(string) The login of the organization.}

\item{team}{(integer or string, optional) The ID or name of the team.}

\item{role}{(string, optional) The role to give the user. For an organization
this is either \code{"member"} or \code{"admin"}, for a team it is either \code{"member"}
or \code{"maintainer"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_memberships()} returns a list of membership properties

\strong{Membership Properties:}
\itemize{
\item \strong{user}: The user login.
\item \strong{organization}: The organization login.
\item \strong{team}: The team name (team membership only).
\item \strong{role}: The role of the user - for organizations it is either \code{"member"}
or \code{"admin"}, for teams it is either \code{"member"} or \code{"maintainer"}.
\item \strong{state}: The state of the membership - either \code{"active"} or \code{"pending"}.
}
}
\description{
This function can be used to invite a user into an organization or team, or
update their role within the organization or team.
}
\details{
Note: you can only invite or update a user if the authenticate user is an
organization "owner" or a team "maintainer".

For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/orgs#set-organization-membership-for-a-user}
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/teams#add-or-update-team-membership-for-a-user}
}
}
\examples{
\dontrun{

  # Invite a new user into an organization
  update_membership("ChadGoymer2", "HairyCoos")

  # Update a user's role in an organization
  update_membership("ChadGoymer2", "HairyCoos", role = "admin")

  # Invite a new user into a team
  update_membership("ChadGoymer2", "HairyCoos", "HeadCoos")

  # Update a user's role in a team
  update_membership(
    user = "ChadGoymer2",
    org  = "HairyCoos",
    team = "HeadCoos",
    role = "maintainer"
  )

}

}
