% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{update_milestone}
\alias{update_milestone}
\title{Update a milestone in a repository}
\usage{
update_milestone(milestone, repo, title, description, due_on, state, ...)
}
\arguments{
\item{milestone}{(string or character) The number or title of the milestone.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{title}{(string, optional) The title of the milestone.}

\item{description}{(string, optional) A description of the milestone.}

\item{due_on}{(string, optional) The milestone due date. This is in the format:
\verb{YYYY-MM-DD HH:MM:SS}.}

\item{state}{(string, optional) The state of the milestone. Either \code{"open"} or \code{"closed"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_milestone()} returns a list of the milestone properties.

\strong{Milestone Properties:}
\itemize{
\item \strong{number}: The number assigned to the milestone within the repository.
\item \strong{title}: The title of the milestone.
\item \strong{description}: The description of the milestone.
\item \strong{state}: The state of the milestone - either "open" or "closed".
\item \strong{open_issues}: The number of open issues within the milestone.
\item \strong{closed_issues}: The number of closed issues within the milestone.
\item \strong{html_url}: The URL of the milestone's web page in GitHub.
\item \strong{creator}: The creator's login.
\item \strong{created_at}: When the milestone was created.
\item \strong{updated_at}: When the milestone was last updated.
\item \strong{due_on}: When the milestone is due.
\item \strong{closed_at}: When the milestone was closed.
}
}
\description{
This function updates a milestone for the specified repository in GitHub. It can be used to
change title, description or due date, but can also be used to close the milestone.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/milestones/#update-a-milestone}
}
}
\examples{
\dontrun{

  # Update the properties of a milestone
  update_milestone(
    milestone   = "test milestone",
    repo        = "ChadGoymer/githapi",
    title       = "updated test milestone",
    description = "This is an updated test milestone",
    due_on      = "2020-12-01")

  # Close a milestone
  update_milestone(
    milestone = "updated test milestone",
    repo      = "ChadGoymer/githapi",
    state     = "closed")

}

}
