% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{update_project}
\alias{update_project}
\title{Update a GitHub project}
\usage{
update_project(
  project,
  name = NULL,
  body = NULL,
  state = NULL,
  permission = NULL,
  private = NULL,
  repo,
  user,
  org,
  ...
)
}
\arguments{
\item{project}{(integer or string) Either the project number or name.}

\item{name}{(string, optional) The new name for the project.}

\item{body}{(string, optional) The new description of the project.}

\item{state}{(string, optional) The new state of the project, either \code{"open"} or \code{"closed"}.}

\item{permission}{(string, optional) The new permissions for the project, either \code{"read"},
\code{"write"}, \code{"admin"} or \code{"none"}.}

\item{private}{(boolean, optional) Whether the project should be private.}

\item{repo}{(string, optional) The repository specified in the format: \code{owner/repo}.}

\item{user}{(string, optional) The login of the user.}

\item{org}{(string, optional) The name of the organization.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_project()} returns a list of the project properties.

\strong{Project Properties:}
\itemize{
\item \strong{id}: The ID of the project.
\item \strong{number}: The number of the project for the repository, user or organization.
\item \strong{name}: The name given to the project.
\item \strong{body}: The description given to the project.
\item \strong{state}: Whether the project is "open" or "closed".
\item \strong{creator}: The user who created the project.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
\item \strong{html_url}: The URL to view the project.
}
}
\description{
This function updates a project in GitHub. It can be used to change the name and body, but
can also be used to close the project or change permissions.
}
\details{
You can update a project associated with either a repository, user or organization, by
supplying them as an input, as long as you have appropriate permissions.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/projects/#update-a-project}
}
}
\examples{
\dontrun{
  # Update the name of a project for a repository
  update_project(
    project = "Repo project",
    name    = "Updated repo project",
    body    = "This is an updated repository's project",
    repo    = "ChadGoymer/test-githapi")

  # Update the state of a project for a user
  update_project(
    name  = "User project",
    state = "closed",
    user  = "ChadGoymer")

  # Update the permissions of a project for an organization
  update_project(
    name       = "Org project",
    permission = "read",
    private    = TRUE,
    org        = "HairyCoos")
}

}
