% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{update_tag}
\alias{update_tag}
\title{Update a tag in a repository}
\usage{
update_tag(tag, ref, repo, force = FALSE, ...)
}
\arguments{
\item{tag}{(string) The name of the tag.}

\item{ref}{(string) Either a SHA, branch or tag used to identify the new commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{force}{(boolean, optional) Whether to force the update if it is not a simple
fast-forward. Default: \code{FALSE}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_tag()} returns a list of the tag properties.

\strong{Tag Properties:}
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{ref}: The full Git reference of the tag.
\item \strong{sha}: The commit SHA the tag is pointing at.
}
}
\description{
This function updates a tag in the specified repository to point at a new commit. It must
be pointed at a commit by providing a Git reference, which can be either a SHA, branch or
tag. For a branch, the head commit is used.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#update-a-reference}
}
}
\examples{
\dontrun{

  update_tag(
    tag  = "new-tag",
    ref  = "6b7b5a090d47fd3ef495620513a3f80da2487b1d",
    repo = "ChadGoymer/githapi")

}

}
