% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-tags.R
\name{update_tags}
\alias{update_tags}
\title{Update tags to new commits.}
\usage{
update_tags(tags, shas, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{tags}{(character) The tag names.}

\item{shas}{(character) The SHAs of the commits to tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the tags, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{ref}: The full git reference for the tag.
\item \strong{url}: The URL to get the tag details from GitHub.
\item \strong{object_sha}: The SHA of the object tagged.
\item \strong{object_type}: The type of the object tagged.
\item \strong{object_url}: The URL to get the object details from GitHub
}
}
\description{
This function updates the specified tags to point at new commits within a repository in
GitHub.
}
\details{
\url{https://developer.github.com/v3/git/refs/#update-a-reference}
}
