% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{update_team}
\alias{update_team}
\title{Update a team's properties in GitHub}
\usage{
update_team(team, name, description, privacy, parent_team, org, ...)
}
\arguments{
\item{team}{(integer or string) The Id or current name of the team.}

\item{name}{(string, optional) The new name of the team.}

\item{description}{(string, optional) The description of the team.}

\item{privacy}{(string, optional) The level of privacy this team should have. Editing teams
without specifying this parameter leaves privacy intact. The options are:

For a non-nested team:
\itemize{
\item \code{"secret"}: only visible to organization owners and members of this team.
\item \code{"closed"}: visible to all members of this organization.
}

For a parent or child team:
\itemize{
\item \code{"closed"}: visible to all members of this organization.
}}

\item{parent_team}{(integer or string, optional) The ID or name of a team to set as the
parent team.}

\item{org}{(string, optional) The organization the team is associated with. Not required
if your are specifying \code{team} and/or \code{parent_team} set by ID.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_team()} returns a list of the team's properties.

\strong{Team Properties:}
\itemize{
\item \strong{id}: The ID of the team.
\item \strong{name}: The name of the team.
\item \strong{organization}: The organization the team is associated with.
\item \strong{slug}: The team slug name.
\item \strong{description}: The description of the team.
\item \strong{privacy}: The privacy setting of the team - either "closed" or "secret".
\item \strong{permission}: The default repository permissions of the team.
\item \strong{parent}: The parent team.
\item \strong{members_count}: The number of members.
\item \strong{repos_count}: The number of repositories the team has access to.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
\item \strong{html_url}: The URL of the team page in GitHub
}
}
\description{
This function updates a team's properties in GitHub. You can only update a team if you have
the appropriate permissions.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/teams/#edit-team}
}
}
\examples{
\dontrun{

  # Update a team
  update_team(
    team        = "TestTeam",
    description = "This is a test team",
    privacy     = "closed",
    parent_team = "TestTeam3",
    org         = "HairyCoos")

}

}
