% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{update_team_repository}
\alias{update_team_repository}
\alias{remove_team_repository}
\title{Update a team's permissions on a repository}
\usage{
update_team_repository(repo, team, org, permission = "pull", ...)

remove_team_repository(repo, team, org, ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{team}{(string) The team name.}

\item{org}{(string) The name of the organization.}

\item{permission}{(string, optional) The permission to set for the team. Either: \code{"pull"},
\code{"push"}, \code{"admin"}, \code{"maintain"} or \code{"triage"}. Default: \code{"pull"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_team_repository()} and \code{remove_team_repository()} returns a TRUE if
successful.
}
\description{
\code{update_team_repository()} allows you to add or update a team's permission on a repository.
\code{remove_team_repository()} removes the team's access to a repository.
}
\details{
The team's permission can be set to:
\itemize{
\item \code{"pull"}: Team members can pull from this repository.
\item \code{"push"}: Team members can pull from and push to this repository.
\item \code{"admin"}: Team members can pull from, push to and administer this repository.
\item \code{"maintain"}: Team members can manage the repository without access to sensitive or
destructive actions. Recommended for project managers. Only applies to repositories owned
by organizations.
\item \code{"triage"}: Team members can proactively manage issues and pull requests without write
access. Recommended for contributors who triage a repository. Only applies to
repositories owned by organizations.
}
}
\examples{
\dontrun{

  # Add read access for the specified team
  update_team_repository(
    repo = "HairyCoos/test-repository",
    team = "test-team",
    org  = "HairyCoos")

  # Update team's permission to "maintain"
  update_team_repository(
    repo       = "HairyCoos/test-repository",
    team       = "test-team",
    org        = "HairyCoos",
    permission = "maintain")

  # Remove team's access
  remove_team_repository(
    repo = "HairyCoos/test-repository",
    team = "test-team",
    org  = "HairyCoos")

}

}
