% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{update_user}
\alias{update_user}
\title{Update your user properties in GitHub}
\usage{
update_user(name, email, blog, company, location, hireable, bio, ...)
}
\arguments{
\item{name}{(string, optional) The new name of the user.}

\item{email}{(string, optional) The publicly visible email address of the
user.}

\item{blog}{(string, optional) The new blog URL of the user.}

\item{company}{(string, optional) The new company of the user.}

\item{location}{(string, optional) The new location of the user.}

\item{hireable}{(boolean, optional) The new hiring availability of the user.}

\item{bio}{(string, optional) The new short biography of the user.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{update_user()} returns a list of the new user properties.

\strong{User Properties:}
\itemize{
\item \strong{id}: The ID of the user.
\item \strong{login}: The login name of the user.
\item \strong{name}: The name of the user.
\item \strong{email}: The public email address of the user.
\item \strong{blog}: The blog address of the user.
\item \strong{company}: The company the user works for.
\item \strong{location}: The location of the user.
\item \strong{hireable}: Whether the user currently hireable.
\item \strong{bio}: The biography of the user.
\item \strong{type}: The type of account.
\item \strong{site_admin}: Whether the user is an administrator.
\item \strong{html_url}: The GitHub page for the user.
}
}
\description{
This function updates your user properties in GitHub. You cannot update
someone else's profile.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/users#update-the-authenticated-user}
}
}
\examples{
\dontrun{

  # Update your name
  update_user(name = "Bob Smith")

  # Update your company
  update_user(company = "Acme")

  # Update your hireable status
  update_user(hireable = TRUE)

}

}
