% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-trees.R
\name{upload_tree}
\alias{upload_tree}
\title{Upload a tree from a directory}
\usage{
upload_tree(
  path,
  repo,
  base_tree,
  ignore = c("\\\\.git", "\\\\.Rproj\\\\.user", "\\\\.Rhistory", "\\\\.RData",
    "\\\\.Ruserdata"),
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{path}{(string) The path to the directory, which is to be uploaded as a tree.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{base_tree}{(string, optional) The SHAs of the tree you want to update with new data.
If not specified, a new tree will be created.}

\item{ignore}{(character) A character vector of regular expressions. If any of these are
detected in a file name they are not uploaded. Default: \code{"\\\\.git"}, \code{"\\\\.Rproj\\\\.user"},
\code{"\\\\.Rhistory"}, \code{"\\\\.RData"} & \code{"\\\\.Ruserdata"}}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the tree, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{sha}: The SHA of the tree.
\item \strong{url}: The URL to get the tree details from GitHub.
}
}
\description{
This function uploads the tree structure from a specified directory into a repository in
GitHub. Files are uploaded as blobs (see \code{\link[=upload_blobs]{upload_blobs()}}) and sub-directories are
uploaded as nested trees \code{\link[=create_tree]{create_tree()}}.
}
\details{
\url{https://developer.github.com/v3/git/trees/#create-a-tree}
}
