% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-blobs-deprecated.R
\name{view_blobs}
\alias{view_blobs}
\title{View information about Git blobs (files)}
\usage{
view_blobs(
  shas,
  repo,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{shas}{(character) The SHAs of the blobs (files).}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the blobs, with the following columns
(see \href{https://developer.github.com/v3/git/blobs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{sha}: The SHA of the blob.
\item \strong{content}: The content of the blob, in base64 encoding.
\item \strong{encoding}: The encoding used for \code{content}. Always \code{"base64"}.
\item \strong{url}: The URL to get the blob details from GitHub.
\item \strong{size}: The size of the blob, in bytes.
}
}
\description{
This function returns details about the blobs (files) for the specified repository in
GitHub.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
