% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branches-deprecated.R
\name{view_branches}
\alias{view_branches}
\title{View information about branches}
\usage{
view_branches(
  branches,
  repo,
  n_max = 1000L,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{branches}{(character, optional) The branch names. If missing or \code{NULL} all the}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.
branches are requested.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the branches, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{name}: The name of the branch.
\item \strong{ref}: The full git reference for the branch.
\item \strong{url}: The URL to get the branch details from GitHub.
\item \strong{object_sha}: The SHA of the object at the HEAD of the branch.
\item \strong{object_type}: The type of the object at the HEAD of the branch.
\item \strong{object_url}: The URL to get the object details from GitHub.
}
}
\description{
This function returns details about the branches for the specified repository in GitHub.
If no branches are supplied then details about all the branches are returned. If branches
are supplied then only information about those ones is returned.
}
\details{
\url{https://developer.github.com/v3/git/refs/#get-a-reference}

\url{https://developer.github.com/v3/git/refs/#get-all-references}
}
