% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{view_columns}
\alias{view_columns}
\alias{view_column}
\title{View columns within a GitHub project}
\usage{
view_columns(project, repo, user, org, n_max = 1000, ...)

view_column(column, project, repo, user, org, ...)
}
\arguments{
\item{project}{(integer or string) Either the project number or name.}

\item{repo}{(string, optional) The repository specified in the format:
\code{owner/repo}.}

\item{user}{(string, optional) The login of the user.}

\item{org}{(string, optional) The name of the organization.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}

\item{column}{(integer or string) The number or name of the column.}
}
\value{
\code{view_columns()} returns a tibble of column properties.
\code{view_column()} returns a list of properties for a single column.

\strong{Column Properties:}
\itemize{
\item \strong{id}: The ID of the column.
\item \strong{name}: The name given to the column.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
\code{view_columns()} summarises columns in a table with the properties as columns
and a row for each column in the project. \code{view_column()} returns a list of
all properties for a single column.
}
\details{
You can summarise all the columns of a project associated with either a
repository, user or organization, by supplying them as an input.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/projects#list-project-columns}
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/projects#get-a-project-column}
}
}
\examples{
\dontrun{

  # View columns in a repository project
  view_columns(
    project = "Test columns",
    repo    = "ChadGoymer/githapi"
  )

  # View columns in a user's project
  view_columns(
    project = "Test columns",
    user    = "ChadGoymer"
  )

  # View columns in an organization's project
  view_column(
    project = "Test columns",
    org     = "HairyCoos"
  )

  # View a column in a repository project
  view_column(
    column  = "Test column",
    project = "Test columns",
    repo    = "ChadGoymer/githapi"
  )

  # View a column in a user's project
  view_column(
    column  = "Test column",
    project = "Test columns",
    user    = "ChadGoymer"
  )

  # View a column in an organization's project
  view_column(
    column  = "Test column",
    project = "Test columns",
    org     = "HairyCoos"
  )

  # View column by ID
  view_column(123456)

}

}
