% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-commits.R
\name{view_commits}
\alias{view_commits}
\title{View information about Git commits.}
\usage{
view_commits(repo, shas, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{shas}{(character) The SHAs of the commits.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the commits, with the following columns
(see \href{https://developer.github.com/v3/repos/commits/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{sha}: The SHA of the commit.
\item \strong{message}: The commit message recorded in GitHub.
\item \strong{author_name}: The author of the commit.
\item \strong{author_email}: The author's email address.
\item \strong{committer_name}: The committer of the commit.
\item \strong{committer_email}: The committer's email address.
\item \strong{date}: The date and time of the commit.
\item \strong{url}: The URL to get the commit details from GitHub.
\item \strong{tree_sha}: The SHA of the git tree.
\item \strong{tree_url}: The URL to get the tree details from GitHub.
\item \strong{parent_sha}: The SHA(s) of the parent commit(s).
\item \strong{parent_url}: The URL(s) to get the parent's details from GitHub.
}
}
\description{
This function returns details about the commits for the specified repository in GitHub.
}
\details{
\url{https://developer.github.com/v3/git/commits/#get-a-commit}
}
