% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{view_files}
\alias{view_files}
\alias{view_file}
\alias{browse_files}
\alias{browse_file}
\title{View files within a repository}
\usage{
view_files(ref, repo, recursive = TRUE, ...)

view_file(path, ref, repo, ...)

browse_files(ref, repo, ...)

browse_file(path, ref, repo, ...)
}
\arguments{
\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{recursive}{(boolean, optional) Whether to list files in subfolders as well.
Default: \code{TRUE}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}

\item{path}{(string) The path to the file, within the repository.}
}
\value{
\code{view_files()} returns a tibble of file properties. \code{view_file()} returns a list
of properties for a single file. \code{browse_files()} and \code{browse_file()} opens the default
browser on the tree or blob page and returns the URL.

\strong{File Properties:}
\itemize{
\item \strong{path}: The path to the file within the repository.
\item \strong{sha}: The SHA of the file blob.
\item \strong{size}: The size of the file in bytes.
\item \strong{html_url}: The URL of the blob's web page in GitHub.
}
}
\description{
\code{view_files()} summarises files in a table with the properties as columns and a row for
each file in the repository. \code{view_file()} returns a list of all properties for a single
file. \code{browse_files()} and \code{browse_file()} open the web page for the commit tree and
blob respectively in the default browser.
}
\details{
You can summarise all the milestones of a repository in a specified \code{state} and change the
order they are returned using \code{sort} and \code{direction}.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository}
}
}
\examples{
\dontrun{

  # View files on the main branch in a repository
  view_files("main", "ChadGoymer/githapi")

  # View properties of a single file in a repository
  view_file(
    path = "README.md",
    ref  = "main",
    repo = "ChadGoymer/githapi")

  # Open the commit's tree page in the default browser
  browse_files("main", "ChadGoymer/githapi")

  # Open the file's blob page in the default browser
  browse_file(
    path = "README.md",
    ref  = "main",
    repo = "ChadGoymer/githapi")

}

}
