% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{view_gists}
\alias{view_gists}
\alias{view_gist}
\alias{browse_gist}
\title{View gists in GitHub}
\usage{
view_gists(user, since, n_max = 1000, ...)

view_gist(gist, ...)

browse_gist(gist, ...)
}
\arguments{
\item{user}{(string, optional) The login of the user. If not specified the authenticated
user is used.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}

\item{gist}{(string) The id of the gist.}
}
\value{
\code{view_gists()} returns a tibble of gist properties. \code{view_gist()}
returns a list of properties for a single gist. \code{browse_gist} opens the
default browser on the gist page and returns the URL.

\strong{Gist Properties:}
\itemize{
\item \strong{id}: The id for the gist
\item \strong{description}: The description of the gist
\item \strong{files}: For \code{view_gists()} just the filenames are returned, for \code{view_gist()} a
tibble of file properties is returned:
\itemize{
\item \strong{filename}: The name of the file
\item \strong{type}: The type of file
\item \strong{content}: The file content
\item \strong{size}: The size of the file in bytes
\item \strong{truncated}: Whether the file content has been truncated
}
\item \strong{owner}: The login if the gist's owner
\item \strong{public}: Whether the gist is public
\item \strong{html_url}: The address of the gist's web page.
\item \strong{created_at}: The time and date the gist was created.
\item \strong{updated_at}: The time and date the gist was last updated.
}
}
\description{
\code{view_gists()} summarises a user's gists in a table with the properties as columns and a
row for each gist. \code{view_gist()} returns a list of all properties for a single gist.
\code{browse_gist()} opens the web page for the gist in the default browser.
}
\details{
When viewing gists, if a user is not specified the gists for the authenticated user are
returned.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/gists/#list-gists-for-a-user}
\item \url{https://developer.github.com/v3/gists/#list-gists-for-the-authenticated-user}
\item \url{https://developer.github.com/v3/gists/#get-a-gist}
}
}
\examples{
\dontrun{

  # View the authenticated user's gists
  view_gists()

  # View a specific user's gists
  view_gists("ChadGoymer")

  # View a specific gist
  view_gist("806dca6b09a39e7b6326a0c8137583e6")

  # Browse a gist
  browse_gist("806dca6b09a39e7b6326a0c8137583e6")

}

}
