% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-commits.R
\name{view_history}
\alias{view_history}
\title{Get information about the history of commits}
\usage{
view_history(
  ref,
  repo,
  n_max = 1000L,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch is
specified the head commit is used. If not specified the default branch is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the commits, with the following columns
(see \href{https://developer.github.com/v3/repos/commits/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{sha}: The SHA of the commit.
\item \strong{message}: The commit message recorded in GitHub.
\item \strong{author_name}: The author of the commit.
\item \strong{author_email}: The author's email address.
\item \strong{committer_name}: The committer of the commit.
\item \strong{committer_email}: The committer's email address.
\item \strong{date}: The date and time of the commit.
\item \strong{url}: The URL to get the commit details from GitHub.
\item \strong{tree_sha}: The SHA of the git tree.
\item \strong{tree_url}: The URL to get the tree details from GitHub.
\item \strong{parent_sha}: The SHA(s) of the parent commit(s).
\item \strong{parent_url}: The URL(s) to get the parent's details from GitHub.
}
}
\description{
This function returns information about all the commits in the history of a specified
commit. The commit can be specified by a branch, tag or SHA.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository}
}
