% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{view_memberships}
\alias{view_memberships}
\alias{view_membership}
\title{View membership of organizations in GitHub}
\usage{
view_memberships(state, n_max = 1000, ...)

view_membership(org, user = NULL, ...)
}
\arguments{
\item{state}{(string, optional) Filter results depending on the \code{state} of the membership.
Can be either \code{"active"} or \code{"pending"}. If not supplied, all memberships are returned for
the authenticated user.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{org}{(string) The login of the organization.}

\item{user}{(string, optional) The login of the user. If \code{NULL} the authenticated user is
used.}
}
\value{
\code{view_memberships()} returns a tibble of membership properties. \code{view_membership()}
returns a list of membership properties for a single organization.

\strong{Membership Properties:}
\itemize{
\item \strong{organization}: The organization login.
\item \strong{user}: The user login.
\item \strong{state}: The state of the membership - either \code{"active"} or \code{"pending"}.
\item \strong{role}: The role of the user in the organization - either \code{"admin"} or \code{"member"}.
}
}
\description{
\code{view_memberships()} summarises the membership of organizations in a table with the
properties as columns and a row for each organization. \code{view_membership()} returns a list
of all membership properties for a single organization.
}
\details{
You can summarise all the organizations a user is a member of by specifying the user login,
or the authenticated user if it is set to \code{NULL}.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/orgs/members/#list-your-organization-memberships}
\item \url{https://developer.github.com/v3/orgs/members/#get-organization-membership}
\item \url{https://developer.github.com/v3/orgs/members/#get-your-organization-membership}
}
}
\examples{
\dontrun{
  # View membership of all organizations the authenticated user is a member of
  view_memberships()

  # View only active memberships for the authenticated user
  view_memberships(state = "active")

  # View the membership of a user in an organization
  view_membership("HairyCoos", "ChadGoymer")

  # View the membership of the authenticated user in an organization
  view_membership("HairyCoos")
}

}
