% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-contents.R
\name{view_readme}
\alias{view_readme}
\title{View the contents of the README file}
\usage{
view_readme(ref, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. If not supplied the head of the default branch is
used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A string containing the README contents.
}
\description{
This function reads the contents of the README file for a repository in GitHub and returns
it as a single string.
}
\details{
\url{https://developer.github.com/v3/repos/contents/#get-the-readme}
}
