% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-releases.R
\name{view_releases}
\alias{view_releases}
\title{View information about a releases}
\usage{
view_releases(repo, tags, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{tags}{(character, optional) The tag names. If missing or \code{NULL} all the releases are
requested. If set to \code{"latest"}, then the latest non-draft, non-prerelease is returned.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the releases, with the following columns
(see \href{https://developer.github.com/v3/repos/releases/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{id}: The unique ID assigned by GitHub.
\item \strong{tag_name}: The name of the tag associated with the release.
\item \strong{name}: The name, or title, of the release.
\item \strong{body}: The body, or description, of the release.
\item \strong{author_login}: The author of the release.
\item \strong{draft}: Whether the release is a draft.
\item \strong{prerelease}: Whether the release is a pre-release.
\item \strong{target_commitish}: The commit or branch the release was applied to.
\item \strong{created_at}: The date and time of creation.
\item \strong{published_at}: The date and time of publication.
\item \strong{assets}: The names of any assets associated with the release.
\item \strong{zipball_url}: The URL to download a zip file of the release commit.
\item \strong{url}: The URL to get the release details from GitHub.
}
}
\description{
This function returns details about the release made in GitHub for the specified repository.
If no tags are supplied then all the releases are returned. If tags are supplied then only
information about those releases is returned. Finally, if "latest" is specified in the place
of a tag the latest non-draft, non-prerelease release is returned.
}
\details{
\url{https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name}

\url{https://developer.github.com/v3/repos/releases/#get-the-latest-release}

\url{https://developer.github.com/v3/repos/releases/#list-releases-for-a-repository}
}
