% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{view_repositories}
\alias{view_repositories}
\alias{view_repository}
\alias{browse_repository}
\title{View repositories for a user or organization}
\usage{
view_repositories(
  user,
  team,
  org,
  sort = "created",
  direction = "desc",
  n_max = 1000,
  ...
)

view_repository(repo, team, org, ...)

browse_repository(repo, ...)
}
\arguments{
\item{user}{(string, optional) The login of the user.}

\item{team}{(string, optional) The team name.}

\item{org}{(string, optional) The name of the organization.}

\item{sort}{(string, optional) The property to order the returned repositories by. Can
be either \code{"created"}, \code{"updated"}, \code{"pushed"} or \code{"full_name"}. Default: \code{"created"}.}

\item{direction}{(string, optional) The direction of the sort. Can be either \code{"asc"} or
\code{"desc"}. Default: \code{"desc"}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}
}
\value{
\code{view_repositories()} returns a tibble of repository properties.
\code{view_repository()} returns a list of properties for a single repository.
\code{browse_repository()} opens the default browser on the repository's page and returns
the URL.

\strong{Repository Properties:}
\itemize{
\item \strong{id}: The ID of the repository.
\item \strong{name}: The name of the repository.
\item \strong{full_name}: The full name of the repository, in the format: \code{owner/repo}.
\item \strong{description}: The description of the repository.
\item \strong{owner}: The owner of the repository.
\item \strong{html_url}: The address of the repository's web page in GitHub.
\item \strong{homepage}: The homepage for the repository.
\item \strong{language}: The dominant programming language in the repository.
\item \strong{size}: The overall size of the repository in bytes.
\item \strong{default_branch}: The name of the default branch.
\item \strong{permission}: The permission the authenticated user has.
\item \strong{private}: Whether the repository is private.
\item \strong{has_issues}: Whether the repository has issues.
\item \strong{has_projects}: Whether the repository has projects.
\item \strong{has_wiki}: Whether the repository has a wiki.
\item \strong{has_pages}: Whether the repository has GitHub Pages.
\item \strong{has_downloads}: Whether the repository has downloads.
\item \strong{allow_squash_merge}: Whether the repository allows squash-merging pull requests.
\item \strong{allow_merge_commit}: Whether the repository allows merging pull requests with a merge
commit.
\item \strong{allow_rebase_merge}: Whether the repository allows rebase-merging pull requests.
\item \strong{fork}: Whether the repository is a fork of another.
\item \strong{archived}: Whether the repository has been archived.
\item \strong{disabled}: Whether the repository has been disabled.
\item \strong{pushed_at}: When the repository was last pushed to.
\item \strong{created_at}: When the repository was created.
\item \strong{updated_at}: When the repository was last updated.
}
}
\description{
\code{view_repositories()} summarises the repositories for a user or organization in a table
with the properties as columns and a row for each repository. \code{view_repository()} returns
a list of a single repository's properties. \code{browse_repository()} opens the web page for
the repository in the default browser.
}
\details{
You can summarise all the repositories associated with either a user or organization, by
supplying them as an input. If neither a user or organization is specified a summary of
the authenticated user's repositories is returned.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/#list-user-repositories}
\item \url{https://developer.github.com/v3/repos/#list-organization-repositories}
\item \url{https://developer.github.com/v3/repos/#list-your-repositories}
}
}
\examples{
\dontrun{

  # View a user's repositories
  view_repositories(user = "ChadGoymer")

  # View an organization's repositories
  view_repositories(org = "HairyCoos")

  # Reorder a user's repositories
  view_repositories(user = "ChadGoymer", sort = "full_name", direction = "asc")

  # View a specific user repository
  view_repository("Test repo", user = "ChadGoymer")

  # View a specific organization repository
  view_repository("Test repo", org = "HairyCoos")

  # Browse a specific user repository
  browse_repository("Test repo", user = "ChadGoymer")

  # Browse a specific organization repository
  browse_repository("Test repo", org = "HairyCoos")

}

}
