% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-commits.R
\name{view_shas}
\alias{view_shas}
\title{Look up the SHA-1 of commit references}
\usage{
view_shas(refs, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{refs}{(character) Git references: either tags or branches. If a branch is specified
the SHA of the head commit is returned.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A named character vector with the supplied refs as names and the SHA as the value.
}
\description{
This function looks up the SHA for the commit associated with the specified git references.
A reference can be either a branch or a tag (or a SHA, but the function will return the
same SHA). If a branch is specified the function returns the SHA of the head commit.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#get-the-sha-1-of-a-commit-reference}
}
