% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{view_statuses}
\alias{view_statuses}
\alias{view_status}
\title{View statuses of a commit within a repository}
\usage{
view_statuses(ref, repo, n_max = 1000, ...)

view_status(ref, repo, ...)
}
\arguments{
\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{view_statuses()} returns a tibble of status properties. \code{view_status()} returns a
string for combined status.

\strong{Status Properties:}
\itemize{
\item \strong{id}: The id of the release.
\item \strong{state}: The state of the status.
\item \strong{description}: The description of the status.
\item \strong{target_url}: The URL linked to the status.
\item \strong{context}: The context of the status.
\item \strong{creator}: The creator of the status.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
\code{view_statuses()} summarises statuses in a table with the properties as columns and a row
for each status for the specified commit in the repository. \code{view_status()} returns the
combined state of all statuses for the commit.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/rest/reference/repos#list-commit-statuses-for-a-reference}
\item \url{https://docs.github.com/en/rest/reference/repos#get-the-combined-status-for-a-specific-reference}
}
}
\examples{
\dontrun{

  # View all statuses for a commit in a repository
  view_statuses("master", "ChadGoymer/githapi")

  # View the combined status
  view_status("master", "ChadGoymer/githapi")

}

}
