% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{view_tag}
\alias{view_tag}
\title{View tags within a repository}
\usage{
view_tag(tag, repo, ...)
}
\arguments{
\item{tag}{(string) The name of the tag.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
\code{view_tags()} returns a tibble of tag properties. \code{view_tag()} returns a list of
properties for a single tag.

\strong{Tag Properties:}
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{ref}: The full Git reference of the tag.
\item \strong{sha}: The commit SHA the tag is pointing at.
}
}
\description{
\code{view_tags()} summarises tags in a table with the properties as columns and a row for each
tag in the repository. \code{view_tag()} returns a list of all properties for a single tag.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/git/refs/#list-matching-references}
\item \url{https://developer.github.com/v3/git/refs/#get-a-single-reference}
}
}
\examples{
\dontrun{
  # View all tags in a repository
  view_tags("ChadGoymer/test-githapi")

  # View a single tag
  view_tag("new-tag", "ChadGoymer/test-githapi")
}

}
