% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{view_tags}
\alias{view_tags}
\alias{view_tag}
\title{View tags within a repository}
\usage{
view_tags(repo, n_max = 1000, ...)

view_tag(tag, repo, ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}

\item{tag}{(string) The name of the tag.}
}
\value{
\code{view_tags()} returns a tibble of tag properties. \code{view_tag()}
returns a list of properties for a single tag.

\strong{Tag Properties:}
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{sha}: The commit SHA the tag is pointing at.
\item \strong{html_url}: The address of the tag's web page in GitHub.
\item \strong{zip_url}: The address to download a zip file of the commit.
\item \strong{tar_url}: The address to download a tar file of the commit.
}
}
\description{
\code{view_tags()} summarises tags in a table with the properties as columns and a
row for each tag in the repository. \code{view_tag()} returns a list of all
properties for a single tag.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://docs.github.com/en/rest/reference/repos#list-repository-tags}
}
}
\examples{
\dontrun{

  # View all tags in a repository
  view_tags("ChadGoymer/githapi")

  # View a single tag
  view_tag("new-tag", "ChadGoymer/githapi")

}

}
