% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{view_teams}
\alias{view_teams}
\alias{view_team}
\alias{browse_team}
\title{View teams in an organization in GitHub}
\usage{
view_teams(org, parent_team, n_max = 1000, ...)

view_team(team, org, ...)

browse_team(team, org, ...)
}
\arguments{
\item{org}{(string, optional) The login of the organization.}

\item{parent_team}{(integer or string, optional) The ID or name of a team. If supplied, all
the child teams will be returned.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{team}{(integer or string) The ID or name of the team.}
}
\value{
\code{view_teams()} returns a tibble of team properties. \code{view_team()} returns a list
of properties for a single team. \code{browse_team()} opens the default browser on the
team's page and returns the URL invisibly.

\strong{Team Properties:}
\itemize{
\item \strong{id}: The ID of the team.
\item \strong{name}: The name of the team.
\item \strong{slug}: The team slug name.
\item \strong{description}: The description of the team.
\item \strong{privacy}: The privacy setting of the team - either "closed" or "secret".
\item \strong{permission}: The default repository permissions of the team.
\item \strong{parent}: The parent team.
\item \strong{html_url}: The URL of the team page in GitHub
}

The following are only returned using \code{view_team()}:
\itemize{
\item \strong{organization}: The organization the team is associated with.
\item \strong{members_count}: The number of members.
\item \strong{repos_count}: The number of repositories the team has access to.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
}
}
\description{
\code{view_teams()} summarises teams in a table with the properties as columns and a row for
each team. \code{view_team()} returns a list of all properties for a single team. \code{browse_team()}
opens the web page for the team in the default browser.
}
\details{
You can summarise all the teams within an organisation, that are children to a parent team
or, if neither are supplied, the teams the authenticated user is a member of.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/teams/#list-teams}
\item \url{https://developer.github.com/v3/teams/#list-child-teams}
\item \url{https://developer.github.com/v3/teams/#list-user-teams}
}
}
\examples{
\dontrun{
  # View teams in an organization
  view_teams("HairyCoos")

  # View a single team
  view_team("HeadCoos", "HairyCoos")

  # Browse a team's GitHub page
  browse_team("HeadCoos", "HairyCoos")
}

}
