% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-trees.R
\name{view_trees}
\alias{view_trees}
\title{View information about Git trees}
\usage{
view_trees(
  shas,
  repo,
  recursive = TRUE,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{shas}{(character) The SHAs of the trees.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{recursive}{(logical) Whether to list files recursively. Default: TRUE.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the trees, with the following columns
(see \href{https://developer.github.com/v3/git/blobs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{tree_sha}: The SHA of the tree.
\item \strong{tree_url}: The URL to get the tree details from GitHub.
\item \strong{path}: The path to the object within the tree.
\item \strong{type}: The type of the object. Either \code{"blob"}, \code{"tree"} or \code{"commit"}.
\item \strong{sha}: The SHA of the object.
\item \strong{size}: The size of the object, in bytes.
\item \strong{url}: The URL to get the object's details from GitHub.
}
}
\description{
This function returns details about the trees of commits for the specified repository in
GitHub. A tree describes the the blobs, or files, usually in a commit.
}
\details{
\url{https://developer.github.com/v3/git/trees/#get-a-tree}
}
