% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-commits.R
\name{compare_files}
\alias{compare_files}
\title{Compare the files of two commits}
\usage{
compare_files(base, head, repo, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{base}{(string) The reference of the base commit.}

\item{head}{(string) The reference of the commit to compare.}

\item{repo}{(string) The repository specified as "owner/repo".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the commits, with the following columns
(see \href{https://developer.github.com/v3/repos/commits/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{filename}: The name of the file.
\item \strong{sha}: The SHA of the file blob.
\item \strong{status}: Whether the file was "added", "deleted" or "changed".
\item \strong{additions}: The number of lines added.
\item \strong{deletions}: The number of lines deleted.
\item \strong{changes}: The number of lines changed.
\item \strong{raw_url}: The URL to the raw file.
\item \strong{contents_url}: The URL to retrieve the contents
\item \strong{patch}: The patch to apply to change the file.
}
}
\description{
This function returns information on all the files that have been changed between two
commits. The commits can be specifed by any git reference, i.e. branch, tag or SHA.
}
\details{
\url{https://developer.github.com/v3/repos/commits/#compare-two-commits}
}
