% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-contents-deprecated.R
\name{download_commit}
\alias{download_commit}
\title{Download the contents of a commit}
\usage{
download_commit(
  path,
  repo,
  ref,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{path}{(string, optional) The path to save to. If not specified the current working
directory is used.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch is
specified the head commit is used. If not specified the default branch is used.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_download_binary]{gh_download_binary()}}.}
}
\value{
A string containing the path. Also, all the files in the commit are downloaded to
the specified location.
}
\description{
This function downloads the entire commit to the specified location. The commit is
downloaded as a zip file, which is then unzipped into the desired location. The commit may
be specified by a branch, tag or SHA.
}
\details{
\url{https://developer.github.com/v3/repos/contents/#get-archive-link}
}
