% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_branch}
\alias{gh_branch}
\title{Get information about a branch}
\usage{
gh_branch(branch, repo, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{branch}{(string) The branch name.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the branch (see GitHub's API documentation for details).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=view_branches]{view_branches()}} instead.
}
\details{
\url{https://developer.github.com/v3/repos/branches/#get-branch}
}
