% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_branches}
\alias{gh_branches}
\title{Get information about all the head commits in each branch.}
\usage{
gh_branches(repo, extended = FALSE, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{extended}{(logical, optional) Whether to return extended information for each branch.
Warning: this may take some time for a lot of branches. Default: \code{FALSE}.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}

\item{limit}{(integer, optional) The maximum number to return. Default: \code{Inf}, which
returns all.}
}
\value{
A tibble describing all the branches (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/branches/#list-branches}
}
\details{
Note: Extended results requires an separate API call for every required branch. This may
impact performance of the function considerably depending on the number of branches.
}
