% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-deprecated.R
\name{gh_collaborators}
\alias{gh_collaborators}
\title{List collaborators}
\usage{
gh_collaborators(
  repo,
  affiliation = NULL,
  n_max = 1000L,
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{affiliation}{(string) Filter collaborators returned by their affiliation.
Can be one of:
\itemize{
\item \code{outside}: All outside collaborators of an organization-owned repository.
\item \code{direct}: All collaborators with permissions to an organization-owned repository,
regardless of organization membership status.
\item \code{all}: All collaborators the authenticated user can see.
\item \code{Default}: all.
}}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the collaborators (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/collaborators/#list-collaborators}
}
