% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{gh_commits}
\alias{gh_commits}
\title{Get information about all the history of commits.}
\usage{
gh_commits(ref, repo, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{ref}{(string) A git reference: either a SHA-1, tag or branch. If a branch is specified
the head commit is used.}

\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{n_max}{(integer) Maximum number to return.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}

\item{limit}{(integer, optional) The maximum number to return. Default: \code{Inf}, which
returns all.}
}
\value{
A tibble describing all the branches (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository}
}
