% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_get}
\alias{gh_get}
\title{Get the contents of a github http request.}
\usage{
gh_get(address, binary = FALSE, accept = "application/vnd.github.raw",
  token = gh_token())
}
\description{
Get the contents of a github http request.
}
