% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gh_gist}
\alias{gh_gist}
\title{Get a single gist}
\usage{
gh_gist(gist, sha, token = gh_token(), api = getOption("github.api"),
  ...)
}
\arguments{
\item{gist}{(string) The ID of the gist in GitHub.}

\item{sha}{(string, optional) The SHA-1 of the version required. Default: latest version.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the gist (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/gists/#get-a-single-gist}
\url{https://developer.github.com/v3/gists/#get-a-specific-revision-of-a-gist}
}
