% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_labels}
\alias{gh_labels}
\title{List all labels for an issue, milestone or repository}
\usage{
gh_labels(repo, issue, milestone, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{issue}{(integer) The number assigned to the issue.}

\item{milestone}{(integer) The number assigned to the milestone.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}
}
\value{
A tibble describing the labels (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/issues/labels/#list-labels-on-an-issue}
\url{https://developer.github.com/v3/issues/labels/#get-labels-for-every-issue-in-a-milestone}
\url{https://developer.github.com/v3/issues/labels/#list-all-labels-for-this-repository}
}
\details{
Note: Must specify either \code{issue} or \code{milestone}, but not both.
}
