% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{gh_membership}
\alias{gh_membership}
\title{Get organization and team membership}
\usage{
gh_membership(user, org, team, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{user}{(string, optional) The GitHub username of the user.}

\item{org}{(string) The name of the organization.}

\item{team}{(integer) The GitHub ID of the team.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list describing the membership (see GitHub's API documentation for details).
}
\description{
\url{https://developer.github.com/v3/orgs/members/#get-organization-membership}
\url{https://developer.github.com/v3/orgs/teams/#get-team-membership}
}
\details{
Note: Must specify either \code{org} or \code{team}, but not both.
}
