% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{gh_page}
\alias{gh_page}
\title{Get multiple pages from the GitHub API}
\usage{
gh_page(
  url,
  n_max = 1000,
  headers = NULL,
  accept = "application/vnd.github.v3+json",
  token = gh_token(),
  proxy = getOption("github.proxy"),
  ...
)
}
\arguments{
\item{url}{(string) The address of the API endpoint.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{headers}{(character, optional) Headers to add to the request. Default: \code{NULL}.}

\item{accept}{(string, optional) The mime format to accept when making the call. Default:
"application/vnd.github.v3+json".}

\item{token}{(string or Token, optional) An authorisation token to include with the
request. If \code{NULL} the OAuth process is triggered. Default: \code{NULL}.}

\item{proxy}{(character, optional) The proxy server to use to connect to the github API.
If \code{NULL} then no proxy is used. Can be set in the option \code{github.proxy} or the
environment variable \code{GITHUB_PROXY}. Default: \code{NULL}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A \code{github} list object consisting of the response, parsed into a list, with the
attributes:
\itemize{
\item \strong{url}: The URLs the request was sent to
\item \strong{request}: The type of HTTP request made
\item \strong{status}: The HTTP status code returned
\item \strong{header}: The HTTP header returned
}
}
\description{
This function is used when requesting a collection of entities. GitHub sets a maximum page
size of 100, so if more are request multiple requests are made and the results are combined.
Each page uses \code{\link[=gh_request]{gh_request()}} to retrieve the contents.
}
\examples{
\dontrun{

  # First 20 users
  gh_page(
    url   = "https://api.github.com/users",
    n_max = 20)

  # First 150 users (two pages)
  gh_page(
    url   = "https://api.github.com/users",
    n_max = 150)

}

}
