% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi.R
\name{gh_page}
\alias{gh_page}
\title{Get and parse the contents of multiple pages from a github http request}
\usage{
gh_page(url, accept = "json", n_max = 1000L, token = gh_token(), ...)
}
\arguments{
\item{url}{(string) URL for GitHub API endpoint.}

\item{accept}{(string) The format of the returned result. Either "json", "raw" or other
GitHub accepted format. Default: "json".}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{...}{Parameters passed to \code{\link[=gh_get]{gh_get()}}.}
}
\value{
A list of the combined parsed responses
}
\description{
Get and parse the contents of multiple pages from a github http request
}
