% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-requests.R
\name{gh_pull_requests}
\alias{gh_pull_requests}
\title{List pull requests}
\usage{
gh_pull_requests(repo, state = NULL, head = NULL, base = NULL,
  sort = NULL, direction = NULL, n_max = 1000L, token = gh_token(),
  api = getOption("github.api"), ...)
}
\arguments{
\item{repo}{(string) The repository specified in the format: \code{"owner/repo"}.}

\item{state}{(string) Either open, closed, or all to filter by state. Default: open}

\item{head}{(string) Filter pulls by head user and branch name in the format of
\code{user:ref-name}. Example: \code{github:new-script-format}.}

\item{base}{(string) Filter pulls by base branch name. Example: gh-pages.}

\item{sort}{(string) What to sort results by. Can be either created, updated, popularity
(comment count) or long-running (age, filtering by pulls updated in the last month).
Default: created}

\item{direction}{(string) The direction of the sort. Can be either asc or desc.
Default: desc when sort is created or sort is not specified, otherwise asc."}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{"GITHUB_TOKEN"} or \code{"GITHUB_PAT"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{"GITHUB_API_URL"} or \code{"https://api.github.com"}.}

\item{...}{Parameters passed to \code{\link{gh_page}}.}
}
\value{
A tibble describing the pull requests (see GitHub's API documentation for details).
}
\description{
url{https://developer.github.com/v3/pulls/#list-pull-requests}
}
