% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_save}
\alias{gh_save}
\title{Download files and save them to a location}
\usage{
gh_save(
  files,
  repo,
  path,
  ref = "master",
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{files}{(string) The paths to the files in the repository.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{path}{(string) The location to save the files to.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. Default: "master".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[=gh_git_tree]{gh_git_tree()}}.}
}
\value{
The file path of the saved file (invisibly).
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=download_files]{download_files()}} instead.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
