% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-data.R
\name{gh_source}
\alias{gh_source}
\title{Source an R file}
\usage{
gh_source(
  file,
  repo,
  ref = "master",
  token = gh_token(),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{file}{(string) The path to the file in the repository}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. Default: "master".}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} or \code{GITHUB_PAT}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}

\item{...}{Parameters passed to \code{\link[base:source]{base::source()}}.}
}
\value{
The result of sourcing the file.
}
\description{
NOTE: This function has been deprecated, please use \code{\link[=source_files]{source_files()}} instead.
}
\details{
\url{https://developer.github.com/v3/git/blobs/#get-a-blob}
}
