% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{gh_token}
\alias{gh_token}
\title{Get a token for accessing GitHub}
\usage{
gh_token(github_token = getOption("github.token"),
  github_oauth_url = getOption("github.oauth_url"),
  githapi_key = getOption("githapi.key"),
  githapi_secret = getOption("githapi.secret"),
  githapi_cache = getOption("githapi.cache"))
}
\arguments{
\item{github_token}{(string, optional) A personal access token. Can be set in the
\code{github.token} option or the \code{GITHUB_TOKEN} environment variable. Default: \code{NULL}.}

\item{github_oauth_url}{(string, optional) The base URL for for the OAuth endpoint. Can
be set in the \code{github.oauth_url} option or the \code{GITHUB_OAUTH_URL} environment variable.
Default: \code{"https://github.com/login/oauth"}.}

\item{githapi_key}{(string, optional) The application ID for accessing githapi. Default:
Set in the \code{githapi.key} option or the \code{GITHAPI_KEY} environment variable.}

\item{githapi_secret}{(string, optional) The secret for the application to access githapi.
Default: Set in the \code{githapi.secret} option or the \code{GITHAPI_SECRET} environment variable.}

\item{githapi_cache}{(boolean or string, optional) The location to store a cached token.
If \code{TRUE} the cache uses the httr default; if \code{FALSE} it does not cache. Can be set
in the \code{"githapi.cache"} option or the \code{GITHAPI_CACHE} environment variable. Default:
\code{FALSE}.}
}
\value{
A token which is either a string, for a personal access token, or a \link[httr:Token]{httr::Token}
object for an OAuth token.
}
\description{
This function initiates the OAuth authentication code grant flow with GitHub. The user
will be redirected to GitHub's login page to enter a user name and password. Once complete
the user can close the window and this function returns a token which can be used for all
calls to the githapi API.
}
\details{
For non-interactive processes a personal access token can be specified, either as an R
option (\code{"github.token"}) or as an environment variable (either \code{GITHUB_TOKEN} or
\code{GITHUB_PAT}).
}
