% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{gh_url}
\alias{gh_url}
\title{Build the URL for the github API}
\usage{
gh_url(..., api = getOption("github.api"))
}
\arguments{
\item{...}{(strings, optional) unnamed strings are built up into a URL path and named
parameters are added as queries.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \code{https://api.github.com}.}
}
\value{
Valid URL (string)
}
\description{
Build the URL for the github API
}
