% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{gh_url}
\alias{gh_url}
\title{Build the URL for the GitHub API}
\usage{
gh_url(..., api = getOption("github.api"))
}
\arguments{
\item{...}{(strings, optional) unnamed strings are built up into a URL path and named
parameters are added as queries.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or \verb{https://api.github.com}.}
}
\value{
Valid URL (string)
}
\description{
This function is used to build the URL for the various endpoints in the GitHub API.
Unnamed strings are used to build the path upon the API and named strings are added as
queries.
}
\examples{
\dontrun{

  # URL for all repositories
  gh_url("repos")

  # URL for the master branch
  gh_url("repos", "ChadGoymer/githapi", "git/refs/heads", "master")

  # URL for a file tree with the recursive option
  gh_url(c("repos", "ChadGoymer/githapi", "git/trees", "234752384"), list(recursive = 1))
}

}
