% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi-deprecated.R
\name{githapi-deprecated}
\alias{githapi-deprecated}
\title{Deprecated functions in package \link{githapi}.}
\description{
The functions listed below are deprecated and will be removed in the near future. When
possible, alternative functions with similar functionality are also mentioned.
}
\section{Git Data Functions}{

\itemize{
\item \code{\link[=gh_git_blob]{gh_git_blob()}}: use \code{\link[=view_blobs]{view_blobs()}} instead.
\item \code{\link[=gh_git_commit]{gh_git_commit()}}: use \code{\link[=view_commits]{view_commits()}} instead.
\item \code{\link[=gh_git_reference]{gh_git_reference()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=gh_git_references]{gh_git_references()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=is_tag]{is_tag()}}: use \code{\link[=tags_exist]{tags_exist()}} instead.
\item \code{\link[=gh_git_tag]{gh_git_tag()}}: is being removed
\item \code{\link[=gh_git_tree]{gh_git_tree()}}: replace with \code{\link[=view_trees]{view_trees()}}
\item \code{\link[=gh_save]{gh_save()}}: replace with \code{\link[=download_files]{download_files()}}
\item \code{\link[=gh_source]{gh_source()}}: replace with \code{\link[=source_files]{source_files()}}
}
}

\section{Repositories Functions}{

\itemize{
\item \code{\link[=gh_tags]{gh_tags()}}: replace with \code{\link[=view_tags]{view_tags()}}
\item \code{\link[=gh_branch]{gh_branch()}}: replace with \code{\link[=view_branches]{view_branches()}}
\item \code{\link[=gh_branches]{gh_branches()}}: replace with \code{\link[=view_branches]{view_branches()}}
\item \code{\link[=is_branch]{is_branch()}}: replace with \code{\link[=branches_exist]{branches_exist()}}
\item \code{\link[=gh_commit]{gh_commit()}}: replace with \code{\link[=view_commits]{view_commits()}}
\item \code{\link[=is_valid_sha]{is_valid_sha()}}: replace with \code{\link[=shas_exist]{shas_exist()}}
\item \code{\link[=gh_commit_sha]{gh_commit_sha()}}: replace with \code{\link[=view_shas]{view_shas()}}
\item \code{\link[=gh_commits]{gh_commits()}}: replace with \code{\link[=view_history]{view_history()}}
\item \code{\link[=gh_compare_commits]{gh_compare_commits()}}: replace with \code{\link[=compare_commits]{compare_commits()}}
\item \code{\link[=gh_compare_files]{gh_compare_files()}}: replace with \code{\link[=compare_files]{compare_files()}}
\item \code{\link[=gh_readme]{gh_readme()}}: replace with \code{\link[=read_files]{read_files()}}
\item \code{\link[=gh_contents]{gh_contents()}}: replace with \code{\link[=read_files]{read_files()}}
\item \code{\link[=gh_download]{gh_download()}}: replace with \code{\link[=download_commit]{download_commit()}}
\item \code{\link[=gh_release]{gh_release()}}: replace with \code{\link[=view_releases]{view_releases()}}
\item \code{\link[=gh_releases]{gh_releases()}}: replace with \code{\link[=view_releases]{view_releases()}}
\item \code{\link[=gh_repository]{gh_repository()}}: replace with \code{\link[=view_repository]{view_repository()}}
\item \code{\link[=gh_repositories]{gh_repositories()}}: replace with \code{\link[=view_repositories]{view_repositories()}}
\item \code{\link[=is_repository]{is_repository()}}: is being removed
}
}

\section{Issues, Pull Requests, labels and milestones}{

\itemize{
\item \code{\link[=gh_label]{gh_label()}}: replace with \code{\link[=view_label]{view_label()}}
\item \code{\link[=gh_labels]{gh_labels()}}: replace with \code{\link[=view_labels]{view_labels()}}
\item \code{\link[=gh_milestone]{gh_milestone()}}: replace with \code{\link[=view_milestone]{view_milestone()}}
\item \code{\link[=gh_milestones]{gh_milestones()}}: replace with \code{\link[=view_milestones]{view_milestones()}}
\item \code{\link[=gh_issue]{gh_issue()}}: replace with \code{\link[=view_issue]{view_issue()}}
\item \code{\link[=gh_issues]{gh_issues()}}: replace with \code{\link[=view_issues]{view_issues()}}
\item \code{\link[=gh_user_issues]{gh_user_issues()}}: replace with \code{\link[=view_issues]{view_issues()}}
\item \code{\link[=gh_assignees]{gh_assignees()}}: is being removed
\item \code{\link[=gh_issue_comments]{gh_issue_comments()}}: replace with \code{\link[=view_comments]{view_comments()}}
\item \code{\link[=gh_issue_comment]{gh_issue_comment()}}: is being removed
\item \code{\link[=gh_event]{gh_event()}}: is being removed
\item \code{\link[=gh_events]{gh_events()}}: is being removed
}
}

\section{Projects Functions}{

\itemize{
\item \code{\link[=gh_project]{gh_project()}}: replace with \code{\link[=view_project]{view_project()}}
\item \code{\link[=gh_projects]{gh_projects()}}: replace with \code{\link[=view_projects]{view_projects()}}
\item \code{\link[=gh_column]{gh_column()}}: replace with \code{\link[=view_column]{view_column()}}
\item \code{\link[=gh_columns]{gh_columns()}}: replace with \code{\link[=view_columns]{view_columns()}}
\item \code{\link[=gh_card]{gh_card()}}: replace with \code{\link[=view_card]{view_card()}}
\item \code{\link[=gh_cards]{gh_cards()}}: replace with \code{\link[=view_cards]{view_cards()}}
}
}

\section{Organizations, teams and users}{

\itemize{
\item \code{\link[=gh_organization]{gh_organization()}}: replace with \code{\link[=view_organization]{view_organization()}}
\item \code{\link[=gh_organizations]{gh_organizations()}}: replace with \code{\link[=view_organizations]{view_organizations()}}
\item \code{\link[=gh_membership]{gh_membership()}}: replace with \code{\link[=view_membership]{view_membership()}}
\item \code{\link[=gh_memberships]{gh_memberships()}}: replace with \code{\link[=view_memberships]{view_memberships()}}
\item \code{\link[=gh_members]{gh_members()}}: replace with \code{\link[=view_users]{view_users()}}
\item \code{\link[=gh_user]{gh_user()}}: replace with \code{\link[=view_user]{view_user()}}
\item \code{\link[=gh_users]{gh_users()}}: replace with \code{\link[=view_users]{view_users()}}
\item \code{\link[=gh_user_email]{gh_user_email()}}: is being removed. Use \code{\link[=view_user]{view_user()}}
\item \code{\link[=gh_team]{gh_team()}}: replace with \code{\link[=view_team]{view_team()}}
\item \code{\link[=gh_teams]{gh_teams()}}: replace with \code{\link[=view_teams]{view_teams()}}
\item \code{\link[=is_member]{is_member()}}: is being removed
\item \code{\link[=is_manager]{is_manager()}}: is being removed
\item \code{\link[=is_collaborator]{is_collaborator()}}: is being removed
\item \code{\link[=gh_collaborators]{gh_collaborators()}}: replace with \code{\link[=view_collaborators]{view_collaborators()}}
\item \code{\link[=gh_permissions]{gh_permissions()}}: replace with \code{\link[=view_collaborator]{view_collaborator()}}
}
}

\section{Pull requests}{

\itemize{
\item \code{\link[=gh_pull_request]{gh_pull_request()}}: replace with \code{\link[=view_pull_request]{view_pull_request()}}
\item \code{\link[=gh_pull_requests]{gh_pull_requests()}}: replace with \code{\link[=view_pull_requests]{view_pull_requests()}}
\item \code{\link[=gh_pull_commits]{gh_pull_commits()}}: replace with \code{\link[=view_pull_request]{view_pull_request()}}
\item \code{\link[=gh_pull_files]{gh_pull_files()}}: replace with \code{\link[=view_pull_request]{view_pull_request()}}
\item \code{\link[=is_pull_merged]{is_pull_merged()}}: is being removed
\item \code{\link[=gh_pull_review]{gh_pull_review()}}: is being removed
\item \code{\link[=gh_pull_reviews]{gh_pull_reviews()}}: replace with \code{\link[=view_pull_request]{view_pull_request()}}
\item \code{\link[=gh_pull_comment]{gh_pull_comment()}}: replace with \code{view_comment()}
\item \code{\link[=gh_pull_comments]{gh_pull_comments()}}: replace with \code{view_comments()}
\item \code{\link[=gh_pull_review_requests]{gh_pull_review_requests()}}: is being removed
}
}

\keyword{internal}
