% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi-deprecated.R
\name{githapi-deprecated}
\alias{githapi-deprecated}
\title{Deprecated functions in package \link{githapi}.}
\description{
The functions listed below are deprecated and will be removed in the near future. When
possible, alternative functions with similar functionality are also mentioned.
}
\section{Git Data Functions}{

\itemize{
\item \code{\link[=gh_git_blob]{gh_git_blob()}}: use \code{\link[=view_blobs]{view_blobs()}} instead.
\item \code{\link[=gh_git_commit]{gh_git_commit()}}: use \code{\link[=view_commits]{view_commits()}} instead.
\item \code{\link[=gh_git_reference]{gh_git_reference()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=gh_git_references]{gh_git_references()}}: use \code{\link[=view_tags]{view_tags()}} or \code{\link[=view_branches]{view_branches()}} instead.
\item \code{\link[=is_tag]{is_tag()}}: use \code{\link[=tags_exist]{tags_exist()}} instead.
\item \code{\link[=gh_git_tag]{gh_git_tag()}}: is being removed
\item \code{\link[=gh_git_tree]{gh_git_tree()}}: replace with \code{\link[=view_trees]{view_trees()}}
\item \code{\link[=gh_save]{gh_save()}}: replace with \code{\link[=download_files]{download_files()}}
\item \code{\link[=gh_source]{gh_source()}}: replace with \code{\link[=source_files]{source_files()}}
}
}

\section{Repositories Functions}{

\itemize{
\item \code{\link[=gh_tags]{gh_tags()}}: replace with \code{\link[=view_tags]{view_tags()}}
\item \code{\link[=gh_branch]{gh_branch()}}: replace with \code{\link[=view_branches]{view_branches()}}
\item \code{\link[=gh_branches]{gh_branches()}}: replace with \code{\link[=view_branches]{view_branches()}}
\item \code{\link[=is_branch]{is_branch()}}: replace with \code{\link[=branches_exist]{branches_exist()}}
\item \code{\link[=gh_commit]{gh_commit()}}: replace with \code{\link[=view_commits]{view_commits()}}
\item \code{\link[=is_valid_sha]{is_valid_sha()}}: replace with \code{\link[=shas_exist]{shas_exist()}}
\item \code{\link[=gh_commit_sha]{gh_commit_sha()}}: replace with \code{\link[=view_shas]{view_shas()}}
\item \code{\link[=gh_commits]{gh_commits()}}: replace with \code{\link[=view_history]{view_history()}}
\item \code{\link[=gh_compare_commits]{gh_compare_commits()}}: replace with \code{\link[=compare_commits]{compare_commits()}}
\item \code{\link[=gh_compare_files]{gh_compare_files()}}: replace with \code{\link[=compare_files]{compare_files()}}
\item \code{\link[=gh_readme]{gh_readme()}}: replace with \code{\link[=read_files]{read_files()}}
\item \code{\link[=gh_contents]{gh_contents()}}: replace with \code{\link[=read_files]{read_files()}}
\item \code{\link[=gh_download]{gh_download()}}: replace with \code{\link[=download_commit]{download_commit()}}
\item \code{\link[=gh_release]{gh_release()}}: replace with \code{\link[=view_releases]{view_releases()}}
\item \code{\link[=gh_releases]{gh_releases()}}: replace with \code{\link[=view_releases]{view_releases()}}
}
}

\section{Projects Functions}{

\itemize{
\item \code{\link[=gh_project]{gh_project()}}: replace with \code{\link[=view_project]{view_project()}}
\item \code{\link[=gh_projects]{gh_projects()}}: replace with \code{\link[=view_projects]{view_projects()}}
\item \code{\link[=gh_column]{gh_column()}}: replace with \code{\link[=view_column]{view_column()}}
\item \code{\link[=gh_columns]{gh_columns()}}: replace with \code{\link[=view_columns]{view_columns()}}
\item \code{\link[=gh_card]{gh_card()}}: replace with \code{\link[=view_card]{view_card()}}
\item \code{\link[=gh_cards]{gh_cards()}}: replace with \code{\link[=view_cards]{view_cards()}}
}
}

\keyword{internal}
