% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githapi-package.R
\docType{package}
\name{githapi}
\alias{githapi}
\title{User-friendly access to the GitHub API for R, consistent with the tidyverse}
\description{
Provides a suite of functions which simplify working with GitHub's API. The functions
have a consistent API and return either lists for single entities and tibbles for
collections.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/ChadGoymer/githapi}
\item Report bugs at \url{https://github.com/ChadGoymer/githapi/issues}
}
}
\author{
Chad Goymer \email{chad.goymer@gmail.com}
}
