% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_repo}
\alias{is_repo}
\title{Checks whether the supplied object is a valid repository name}
\usage{
is_repo(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a valid repository name, FALSE otherwise
}
\description{
A valid repository name is comprised of two strings separated by a "/".
}
