% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_sha}
\alias{is_sha}
\title{Checks whether the supplied object is a valid SHA}
\usage{
is_sha(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a valid SHA, FALSE otherwise
}
\description{
A valid SHA is 40 characters long and only contains the characters 0-9 & a-f.
}
