% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{view_commits}
\alias{view_commits}
\alias{view_commit}
\alias{browse_commits}
\alias{browse_commit}
\title{View commits within a repository}
\usage{
view_commits(ref, repo, path, author, since, until, n_max = 1000, ...)

view_commit(ref, repo, ...)

browse_commits(ref, repo, ...)

browse_commit(ref, repo, ...)
}
\arguments{
\item{ref}{(string) Either a SHA, branch or tag used to identify the commit.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{path}{(string, optional) Only commits containing this file path will be returned.}

\item{author}{(string, optional) Author login to filter commits by.}

\item{since}{(string, optional) A date & time to filter by. Must be in the format:
\verb{YYYY-MM-DD HH:MM:SS}.}

\item{until}{(string, optional) A date & time to filter by. Must be in the format:
\verb{YYYY-MM-DD HH:MM:SS}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}} or \code{\link[=gh_request]{gh_request()}}.}
}
\value{
\code{view_commits()} returns a tibble of commit properties. \code{view_commit()} returns
a list of properties for a single commit. \code{browse_commits()} and \code{browse_commit} opens
the default browser on the commit's history or details page and returns the URL.

\strong{Commit Properties:}
\itemize{
\item \strong{sha}: The commit SHA.
\item \strong{message}: The commit message.
\item \strong{author_login}: The author's account login.
\item \strong{author_name}: The author's name.
\item \strong{author_email}: The author's email address.
\item \strong{committer_login}: The committer's account login.
\item \strong{committer_name}: The committer's name.
\item \strong{committer_email}: The committer's email address.
\item \strong{tree_sha}: The SHA of the commit's file tree.
\item \strong{parents}: The SHAs of the parent commits.
\item \strong{date}: The time and date the commit was made.
\item \strong{html_url}: The address of the commit's web page.
}
}
\description{
\code{view_commits()} summarises commits in a table with the properties as columns and a row for
each commit in the history of the given reference. \code{view_commit()} returns a list of all
properties for a single commit. \code{browse_commits()} and \code{browse_commit()} opens the web page
for the commit history and commit details respectively in the default browser.
}
\details{
For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository}
\item \url{https://developer.github.com/v3/repos/commits/#get-a-single-commit}
}
}
\examples{
\dontrun{

  # View the history of commits for the main branch
  view_commits("main", "ChadGoymer/githapi")

  # View commits where the README.md file has been changed
  view_commits(
    ref  = "main",
    repo = "ChadGoymer/githapi",
    path = "README.md")

  # View commits created by an author
  view_commits(
    ref    = "main",
    repo   = "ChadGoymer/githapi",
    author = "ChadGoymer")

  # View commits within a time window
  view_commits(
    ref   = "main",
    repo  = "ChadGoymer/githapi",
    since = "2020-01-01 00:00:00",
    until = "2020-04-01 00:00:00")

  # View the properties of the last commit on the main branch
  view_commit("main", "ChadGoymer/githapi")

  # View the properties of the commit with tag "0.8.7"
  view_commit("0.8.7", "ChadGoymer/githapi")

}

}
