% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories-contents.R
\name{view_files}
\alias{view_files}
\title{View information about a files}
\usage{
view_files(paths, repo, ref, token = getOption("github.token"),
  api = getOption("github.api"), ...)
}
\arguments{
\item{paths}{(string, optional) The paths to the files or directories in the repository.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{ref}{(string, optional) A git reference: either a SHA-1, tag or branch. If a branch
is specified the head commit is used. If not specified the head of the default branch is used.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the files, with the following columns
(see \href{https://developer.github.com/v3/repos/contents/}{GitHub's documentation}
for more details):
\itemize{
\item \strong{name}: The name of the file.
\item \strong{path}: The path to the file.
\item \strong{sha}: The SHA of the file.
\item \strong{size}: The size of the file in bytes.
\item \strong{type}: Either \code{"file"}, \code{"dir"}, \code{"symlink"} or \code{"submodule"}.
\item \strong{url}: The URL to the file contents.
\item \strong{html_url}: The URL to view the file in GitHub.
\item \strong{git_url}: The URL to the Git blob.
\item \strong{download_url}: The URL to download the file.
}
}
\description{
This function returns details about the files in GitHub for the specified repository.
If no files are supplied then all the files in the reference commit are returned. If files
are supplied then only information about those files is returned. Finally, if a directory
is specified then details about all the files within the directory is returned.
}
\details{
\url{https://developer.github.com/v3/repos/contents/#get-contents}
}
