% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{view_memberships}
\alias{view_memberships}
\alias{view_membership}
\title{View membership of organizations or teams}
\usage{
view_memberships(state, n_max = 1000, ...)

view_membership(user, org, team, ...)
}
\arguments{
\item{state}{(string, optional) Filter results depending on the \code{state} of the membership.
Can be either \code{"active"} or \code{"pending"}. If not supplied, all memberships are returned for
the authenticated user.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{user}{(string) The login of the user.}

\item{org}{(string) The login of the organization.}

\item{team}{(integer or string, optional) The ID or name of the team.}
}
\value{
\code{view_memberships()} returns a tibble of membership properties. \code{view_membership()}
returns a list of membership properties for a single organization or team.

\strong{Membership Properties:}
\itemize{
\item \strong{user}: The user login.
\item \strong{organization}: The organization login.
\item \strong{team}: The team name (team membership only).
\item \strong{role}: The role of the user - for organizations it is either \code{"member"} or \code{"admin"},
for teams it is either \code{"member"} or \code{"maintainer"}.
\item \strong{state}: The state of the membership - either \code{"active"} or \code{"pending"}.
}
}
\description{
\code{view_memberships()} summarises the authenticate user's membership in organizations in
a table with the properties as columns and a row for each organization. \code{view_membership()}
returns a list of all membership properties for a user in a single organization or team.
}
\details{
Note: you can only view another user's membership in an organization or team if the
authenticate user is also a member.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/orgs/members/#list-your-organization-memberships}
\item \url{https://developer.github.com/v3/orgs/members/#get-organization-membership}
\item \url{https://developer.github.com/v3/teams/members/#get-team-membership}
}
}
\examples{
\dontrun{

  # View membership of all organizations the authenticated user is a member of
  view_memberships()

  # View only active memberships for the authenticated user
  view_memberships(state = "active")

  # View the membership of a user in an organization
  view_membership("ChadGoymer", "HairyCoos")

  # View the membership of a user in a team
  view_membership(user = "ChadGoymer", org = "HairyCoos", team = "HeadCoos")

}

}
