% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{view_projects}
\alias{view_projects}
\alias{view_project}
\alias{browse_project}
\title{View GitHub projects}
\usage{
view_projects(repo, user, org, state = "open", n_max = 1000, ...)

view_project(project, repo, user, org, ...)

browse_project(project, repo, user, org, ...)
}
\arguments{
\item{repo}{(string, optional) The repository specified in the format: \code{owner/repo}.}

\item{user}{(string, optional) The login of the user.}

\item{org}{(string, optional) The name of the organization.}

\item{state}{(string, optional) Indicates the state of the projects to return. Can be
either "open", "closed", or "all". Default: \code{"open"}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{project}{(integer or string) The number or name of the project.}
}
\value{
\code{view_projects()} returns a tibble of project properties. \code{view_project()}
returns a list of properties for a single project. \code{browse_project()} opens the default
browser on the prject page and returns the URL invisibly.

\strong{Project Properties:}
\itemize{
\item \strong{id}: The ID of the project.
\item \strong{number}: The number of the project for the repository, user or organisation.
\item \strong{name}: The name given to the project.
\item \strong{body}: The description given to the project.
\item \strong{state}: Whether the project is "open" or "closed".
\item \strong{creator}: The user who created the project.
\item \strong{created_at}: When it was created.
\item \strong{updated_at}: When it was last updated.
\item \strong{html_url}: The URL to view the project.
}
}
\description{
\code{view_projects()} summarises projects in a table with the properties as columns and a row
for each project. \code{view_project()} returns a list of all properties for a single job.
\code{browse_project()} opens the web page for the project in the default browser.
}
\details{
You can summarise all the projects associated with either a repository, user or
organisation, by supplying them as an input.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/projects/#list-repository-projects}
\item \url{https://developer.github.com/v3/projects/#list-user-projects}
\item \url{https://developer.github.com/v3/projects/#list-organization-projects}
\item \url{https://developer.github.com/v3/projects/#get-a-project}
}
}
\examples{
\dontrun{
  # View a repository's projects
  view_projects("ChadGoymer/githapi")

  # View a user's projects
  view_projects(user = "ChadGoymer")

  # View an organisation's projects
  view_projects(org = "Tidyverse")

  # View closed projects
  view_projects("ChadGoymer/githapi", state = "closed")

  # View all projects
  view_projects("ChadGoymer/githapi", state = "all")

  # View a specific project
  view_project("Prioritisation", "ChadGoymer/githapi")

  # Browse a specific project
  browse_project("Prioritisation", "ChadGoymer/githapi")
}

}
