% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-tags.R
\name{view_tags}
\alias{view_tags}
\title{View information about a tags}
\usage{
view_tags(
  tags,
  repo,
  n_max = 1000L,
  token = getOption("github.token"),
  api = getOption("github.api"),
  ...
)
}
\arguments{
\item{tags}{(character, optional) The tag names. If missing or \code{NULL} all the tags are
requested.}

\item{repo}{(string) The repository specified in the format: \code{owner/repo}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: 1000.}

\item{token}{(string, optional) The personal access token for GitHub authorisation. Default:
value stored in the environment variable \code{GITHUB_TOKEN} (or \code{GITHUB_PAT}) or in the
R option \code{"github.token"}.}

\item{api}{(string, optional) The URL of GitHub's API. Default: the value stored in the
environment variable \code{GITHUB_API} or in the R option \code{"github.api"}.}

\item{...}{Parameters passed to \code{\link[=gh_request]{gh_request()}}.}
}
\value{
A tibble describing the tags, with the following columns
(see \href{https://developer.github.com/v3/git/refs/}{GitHub's documentation} for more details):
\itemize{
\item \strong{name}: The name of the tag.
\item \strong{ref}: The full git reference for the tag.
\item \strong{url}: The URL to get the tag details from GitHub.
\item \strong{object_sha}: The SHA of the object tagged.
\item \strong{object_type}: The type of the object tagged.
\item \strong{object_url}: The URL to get the object details from GitHub
}
}
\description{
This function returns details about the tags for the specified repository in GitHub. If no
tags are supplied then details about all the tags are returned. If tags are supplied then only
information about those ones is returned.
}
\details{
\url{https://developer.github.com/v3/git/refs/#get-a-reference}

\url{https://developer.github.com/v3/git/refs/#get-all-references}
}
