% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{view_users}
\alias{view_users}
\alias{view_user}
\alias{browse_user}
\title{View users in GitHub}
\usage{
view_users(org, team, role = "all", n_max = 1000, ...)

view_user(user, ...)

browse_user(user, ...)
}
\arguments{
\item{org}{(string, optional) The name of the organization.}

\item{team}{(string, optional) The name of the team.}

\item{role}{(string, optional) Filter the result by role. Can specify either \code{"admin"},
\code{"member"} or \code{"all"}. Default: \code{"all"}.}

\item{n_max}{(integer, optional) Maximum number to return. Default: \code{1000}.}

\item{...}{Parameters passed to \code{\link[=gh_page]{gh_page()}}.}

\item{user}{(integer) The login of the user.}
}
\value{
\code{view_users()} returns a tibble of user properties. \code{view_user()}
returns a list of properties for a single user.  \code{browse_user()} opens the default
browser on the user's page and returns the URL invisibly.

\strong{user Properties:}
\itemize{
\item \strong{id}: The ID of the user.
\item \strong{login}: The login name of the user.
\item \strong{name}: The name of the user (only available in \code{view_user()}).
\item \strong{email}: The public email address of the user (only available in \code{view_user()}).
\item \strong{blog}: The blog address of the user (only available in \code{view_user()}).
\item \strong{company}: The company the user works for (only available in \code{view_user()}).
\item \strong{location}: The location of the user (only available in \code{view_user()}).
\item \strong{hireable}: Whether the user currently hireable (only available in \code{view_user()}).
\item \strong{bio}: The biography of the user (only available in \code{view_user()}).
\item \strong{type}: The type of account.
\item \strong{site_admin}: Whether the user is an administrator.
\item \strong{html_url}: The GitHub page for the user.
}
}
\description{
\code{view_users()} summarises users in a table with the properties as columns and a row
for each user. \code{view_user()} returns a list of all properties for a single user.
\code{browse_user()} opens the web page for the user in the default browser.
}
\details{
You can summarise all the users associated with either an organization or a team within
an organization. If none of those are supplied the first \code{n_max} users of GitHub are
returned.

For more details see the GitHub API documentation:
\itemize{
\item \url{https://developer.github.com/v3/orgs/members/#members-list}
\item \url{https://developer.github.com/v3/teams/members/#list-team-members}
\item \url{https://developer.github.com/v3/users/#get-all-users}
\item \url{https://developer.github.com/v3/users/#get-a-single-user}
}
}
\examples{
\dontrun{
  # View users collaborating on a repository
  view_users(repo = "ChadGoymer/githapi")

  # View users in an organization
  view_users(org = "HairyCoos")

  # View users in a team within an organization
  view_users(org = "HairyCoos", team = "HeadCoos")

  # View the admins of an organization
  view_users(org = "HairyCoos", role = "admin")

  # View a single user
  view_user("ChadGoymer")

  # Browse a user's GitHub page
  browse_user("ChadGoymer")
}

}
